Added a "continue move" action.
There are four parts to this patch:

1) add the action to the usual places 2) add a gamemap::location to class unit,
which stores the location we were trying to move to, 3) clear this
location on end-of-turn, or if the unit moves again this turn, and 4)
if a move is interrupted, reslect the unit's hex.

The new hotkey is 't'. To use this feature, select a unit that had its
move interrupted by seeing allies/enemies, and right click >> "Continue
Move", or press the 't' key.