Fixed a problem where the AI would get stuck...
...and never finish making its move. This problem was easily
reproducible on some scenarios, such as e.g.  the first turn of Hasty
Alliance.

Developer info: ai::get_villages was refactored from version 0.7.6 to 0.7.7.
One of the changes caused its return value to be dependent on whether the new
function ai::get_village_combinations(...) finds any moves. However, some of
these moves are not necessarily valid because the destination is taken by
another unit. I'm not sure the cause of that. In any case, the resolution is
for get_villages(...) to only return "true" if it actually makes any moves,
ignoring the invalid ones. This is accomplished with a simple "moves_made"
variable.

TODO: look into get_village_combinations(...) and figure out why we're trying
to take a village that already has a unit on it.