One of the most frightening bug I ever encountered:
it was allowing units to walk on the water on some computers. And I
don't even want to think about all the other problems it could have
caused: potentially any floating-point value ever read was wrong. It
will annoy people who are expecting the damage EV to be displayed with
a comma, but I consider it an acceptable loss given the extent of the
issue it fixes. Let's hope it is sufficient for everybody.