Now the preprocessor output is also a stream, and so is the parser input, ...
...hence they can be plugged, yay. Also introduce a scoped_istream to
prevent leaking descriptors in case of exceptions. Once again, the
stream changes are purely on the interface level for the time being,
so there can be a slight performance hit when no cache is present. But
once the cache is built, everything should be fine, even faster than
it was before the changes. And isn't that the point of a cache?