Fix corrupted translations...
...due to the xgettext comment gone too far away from the translatable
string. The value '125' is also wrong, it should simply be '100' since
the gold was already scaled. And even '100' is wrong since it is a
settable WML field. And as such the string was wrong too, so remove
everything. As a consequence, the old problem is still not fixed, but
translators can at least do their work now.