Speed up the pathfinder by using a heap for the open list...
...instead of a sorted vector. Although the patch is big, it is mostly
identation cleanup (death to the 200 char long lines). The only code
changes are: 1. removal of a_star_stable_sort_and_merge since it is no
more a sorted vector, 2. modification of a_star_explore_neighbours so
that it restores the heap structure on return, 3. trivial changes in
a_star_search so that it pops the heap. Shortcoming: in order for the
patch to stay trivial, a_star_explore_neighbours makes the heap anew
in case of node modification, instead of doing an incremental rebuild.