Review of the A* allocation code.
A lot of noise due to switching to Wesnoth coding style: underscore
after the private members. The memory leaks on destruction are now
plugged, and the node pool is now private to the node world; the
pathfinder doesn't know about it anymore. The parano assertions are no
longer needed. All these changes are only interface changes, the
implementation is not changed: this is still the same page-based
allocation strategy.