Fix the issue with extra directory level being added to campaigns on download.
campaign_name is now set to the campaign_name and name is cleared (in
the event that a bad client were to set its value).  The issue is that
the top level is treated as a directory entry, but the clients make a
special check for a null name. So if the name isn't null it is treated
as a directory name.

However there is still a security issue that I will be addressing in a
separate patch coming shortly.

The [dir] and [file] "name" attributes to not appear to be checked on either
upload or download and should really be checked in both places. It looks like
"name" can be checked to make sure it is not null and then use the same
valid filename check that is applied the campaign name.