mkdos, a script to create DOS versions...
...of all the text files distributed with the game: the various
translations of the MANUAL, the changelog, the license information in
COPYING, the instructions in INSTALL, and the README.  Changes
sequences of lines to DOS-style single paragraphs, and attempts to
preserve displayed text and tables.  The resulting files should be
ready to be shipped with the Windows package, and for most people
these files will be shown in NotePad when double-clicked.