Neaten display::set_paths() and display::set_reach_map().
Bug #5145 was caused by a missing set_paths(NULL), because the display
object used to keep the paths pointer it was passed.

This is unnecessary, because the gui only needs to keep which tiles to
highlight: it does not need to keep the paths themselves.  Also, both
interfaces are very similar and should be combined.  While we're
there, rename them: we don't use it for the actual "route"
highlighting (ie. the footsteps), just for showing where a unit/units
can reach.