Separate "cures" from "heals" cleanly:
...remove global assumptions about cure/heal amounts, too.

I also added an explicit number to how much terrain heals (it's still
assumed to cure).

I also separated "cures" and "heals" from the user's perspective,
which unfortunately means the description of "heals" does not include
the number of HP any more.  It would be nice if the tooltip could use
variables...