Fix bug found by Soliton,
...where damage screen showed double attacker damage when defender had
charge.  In particular, two thieves on either side of a chocobone at
night claimed to be able to do 22-3.

This is a hack: the real fix is to separate all the mutables out of
attack_type and create an attack_context which is constructed with
info about both attacks (currently we hand the opponent's attack to
set_specials_context before we've called set_specials_context on that
opponent attack).