Change the ordering of the Abilities and Traits Help sections...
...so that the overviews come first.

Since Traits is done just in help.cfg, I turned off sorting. The
traits are now sorted by english name, but grouped by normal traits
and special traits.  To take care of Abilities, I made a change so
that "generated" topics always get sorted. For nonsorted sections, the
topics listed in the config file come first in that order and are then
followed by the sorted list of generated topics. I also turned off
sorting for the Abilities section, so that the Overview topic would be
first.