Fix units not being shown when moves done, and glitch when teleporting
(drawn in wrong place).

The correct way of moving a unit is: set_hidden(true),
unit_display::move(), set_hidden(false).  This means that unit count
on the status display does not drop mysteriously, unlike doing the
obvious: units.extract(loc), unit_display::move().  But it means that
unit_display needs to use a copy for the temporary unit it draws on
the screen, and mark that copy non-hidden, otherwise it draws a hidden
unit, ie. nothing.