Sapient pointed out that we un-mute in lots of places, such as preferences.
This feature seems unloved, as this bug has been here for a while.
Plus, it'd be confusing to mute and then go to prefs and see that
music is on, volume is high, but you can't hear it...

Simplify it: actually toggle music and sound on/off, rather than just
temporarilty overriding volume.