Give undead units the undead trait...
...even when they wouldn't normally get traits such as leaders in
multiplayer games.

This is just a temporary hack. Since undead is really applied to a unit
type rather than single units, it should be an ability and not a trait.
While its a bit late for 1.2, what should probably happen is that undead
should be dropped and not living should be changed to an ability rather than
a status and all of the undead unit types (plus the miscellanious not living
campaign units) should get the not living ability. Undead types already get
the not_living status and that is what protects them from poison and plague.
The undead trait is used to indicate that, put doesn't really provide the
protection.