Fix speed on Wesbench_ai:
introduce an AI cache which stores previous calculations on what
weapon to use against an opponent.

This operates on the assumption that similar types will usually use
the same weapon, and is about 99% accurate.  Worst case, AI
underestimates the damage it could do.

Profiling shows that (after the next checking which fastpaths simple
attack_prediction) we were spending ~30% of time figuring out what the
effects of weapons should be, including 10% of time in timeofday_at.

This could be significantly cut if we eliminate unused generality in
all these calculations.

Time for wesbench before (-O3 -march=pentium-m): ~14 seconds.  Time
after: 2.3 seconds.