The removal of extra |s in interpolated strings is a mistake.
This makes it hard to put a | in a string and don't think removing
them to prevent confusion really has that effect.

I also changed it so that if the variable name is empty that the replacement
string will be a $. This makes it easy to put $s in interpolated strings
and the previous behavior was useless.