Fixed transitions of the custom lava in UtBS...
(it now uses the old lava transitions, since the chasm edges would
have made no sense). 

Also made it to be an alias of cave and lava (best movement, worst
defense), which lowers at least elves' movement costs on it from 3 to
2, which shouldn't matter too much (reason being that displaying it as
an alias of snow looks rather silly, this could be changed back too if
needs be).