Elyssa's special units go away.
The only effects on gameplay will be that (a) Arch Mage Elyssa has the
new archmage HP of 54 rather than the old 57.  and (b) her portait
doesn't switch to the silver one if she advances to Silver Mage.

The former is arguably a bug fix.  The latter will get fixed when Sapient
writes the extension for events bound to single-unit declarations.  In the
meantime, we're dumping a bunch of cruft and bloat.