Change the Windows port to use either the native compiler support...
...for the POSIX dirent(3) calls (under GNU C and Borland C) or an
internal emulation of same from the BSD-licensed unixem library
(Intel, Metrowerks, Watcom, MSC). This was patch #735.

This patch has been tested and is known to work under GNU C for
Windows, using its native dirent(3) support.  The emulation lifted from
the unixem library (http://synesis.com.au/software/unixem.html) is not
yet tested, but should be trivial to debug if it has any problems.

The actual point of this change was to unsnarl the logic of
get_files_in_dir() so as to be able to implement special file 
handling without worrying about the underlying OS.