Previously (in 1.3.3) the attack cursor was used even when no attacks left
Now use normal cursor in this case and also when trying to move on a
friendly unit (that one was new).

Continue to move checks in the function to check if an attack is valid
for fixing this So do again the simplification reverted in 2007-06-02T15:59:52Z!koraq@xs4all.nl
because now it works.