Add code and data for a new game_config key : "defense_color_scale"
Currently it will be used only for the movement hint, but I choose a
general name, in case of other uses.

There is 11 colors, each one for each 10 intervals and the last one
for the maximum.

I kept my previous colors: pure yellow for 50% and subtract multiple
of 50 in red or green for other values.