fix a small error in the function to delete tags of a string,
...was buggy when starting with numbers.

This fix a tiny bug : player name starting with a digit, had this
digit repeated twice in the satus table

But i will use more this function in my next commit, so it must be
cleaner.  The function itself is very ugly and has a clone in
utils.cpp. I will clean this mess after the release.