Merge 2007-07-21T14:55:13Z!koraq@xs4all.nl from Mordante:
the unit name generation could with different locales call get_random() a
different number of times. This lead to different names and traits.

Changed to call random a fixed number of times which fixes the traits.
The names are still wrong and still cause OOS errors but these should
not be fatal (Fixes bug #9472 partly)