Don't use the attack/move cursor and attack direction indicator when observing.
This also help to see that it's not your turn and that you can't send
move/attack commands.  This fix bug #9610 but it can maybe still exist
after some events during your turn.  The complete fix need a
mouse-code more up-to-date with the game (which will need more code
changes)