Starting to move fog/shroud later in the rendering sequence of a tile.
The current change render it on top of grid and unreachable mask ("the
black stripes"). Two good points

- We can now see the fog limit even with the the black stripes (was hard before)

- no grid on the shroud (what's the point?) and it is less ugly when
there is fog.  Previously the fog "blurred" terrains but not grid,
which was too much visible compared to unfogged tiles.

The final aim of this is to draw fog/shroud on top of units