Add a second window to the "Status Table"
...which shows the basic game settings for each player. Unfortunatly
the current dialog implementation seems to allow only 2 functional
buttons together with a (menu)list, so I had to move the "Ok" button
which allows you to scroll to the selected leader to the second
screen.

I hope this feature is not used too often, so it's ok there ...

I have mostly tested the "Game Settings" display with multiplayer games, please
test it with scenarios which have more special WML settings.