This *should* be a refactoring step that simplifies the end-of-scenario logic
...without changing behavior; it just moves some stuff, including
autosaving, out of a try/catch scope, and as far as I can tell that
stuff never throws the relevant exceptions.

But if strange bugs start showing up at end of scenario, this would
be a good place to bisect while running regressions.

This is the first commit in an attempted fix for bug #9894.