Add an "Animate Map" switch in advanced preferences.
(temporary location) I wait to see where other future animation switch will go

This stops flags and animated terrains in the first frame, for
simplicity and because the low use of animation prevent to test what
is the best solution

But this choice also allow to choose the default frame.
e.g. a geyser terrain may want to always show the active (or sleeping) state