Add a "Races" section, with one topic for each race.
The topic contains the race description if defined and a list of links
to units of this race. Unit's page have also a field "race:" with a
link to these pages.

Not sure if it will stay like this, but now the code to handle it is
ready.

But note that it's now very similar to Abilities and Specials
sections.