Finish to make the use of update_last_draw_time() consistent.
We update it before get the current frame, because if the redraw is
trigerred by a need_update, we need to draw the new frame not the old
one.

I think that the previous use just caused a tiny lag when we redraw
every x ms, but can cause a "duration of the frame" delay if we redraw
only when an update is needed (which is the good optimal
behavior). This also can prevent skipping of the last frame (if after
drawing, we update a new frame but finish the anim) PS: other game
objects already use update_last_draw_time() like that.


Former-commit-id: f7d443af8b06858781e612c6b1222c3ed67036de