Load unit animations data only when the first unit of a type is created.
Previously, we directly loaded all the images of all unit types.

Here it makes the initial loading of units few times faster (few sec
gain) and spare some memory. But it also helps when loading MP,
Campaigns, add-ons...

The only drawback is when discovering a lot of various units at the
same time, but it is supposed to happen only when loading or starting
a game, so seems ok to load data there, and it only happens the first
time


Former-commit-id: b015eb8dacc008bd9984c67311c37a73245ec270