Speed-up the validation of the cache by skipping media directories...
...(images and sounds) because we don't cache those. This seems to
really help this process.  (~80% of files are images and the check is
proportional to the number of files) The only problem is if weird UMC
put some cfg in /images. But the unchecked cache will just annoy the
author when he will editing only this cfg.  This will also helps
artists by avoiding useless cache rebuilding.  PS: also did a tiny
safe change in Mac-specific code for unicode filename, and a cleaner
file extension check is much less effective than the directory skip.


Former-commit-id: 1342bd003cb8b9311530946e015eac1fa03113fd