Experimental middle button mouse scrolling:
Keep the button pressed and the direction and speed of the scrolling
will depends to the position of the cursor relatively to the center of
the view.

It's allow slow and precise short scrolling, but also long fast ones.

The speed must be tuned but already use the scroll speed of
preferences.  (currently top-speed is equal to direction-keys and
mouse-at-border)

PS: it also replace the old middle-click behavior, since just do a
long click has now a similar effect.

Former-commit-id: 71e0b8f8dedf795f0748b4722754fe2cc7202fdf