this change in 2008-01-24T20:52:29Z!benoit.timbert@free.fr appears to be incorrect so I am reverting it.
if there is a reason why this was changed from an entity_location to a
plain gamemap::location (which disallows the intelligent filtering
that prevents events from firing for disappeared units) then please
provide that reason in the code comments or in the commit log. Thanks.

I have not tested to see if this reintroduces bug #10801.


Former-commit-id: 0f2037e3e75c5195f81e44ae1b2f87ceb58f45ce