Address bug #10885.
This actually seems to be a problem in the theme definition, not the
underlying display engine -- the theme defines a fixed-sized report
area that assumes position numbers will always be at most two digits.
We can live with not fixing it, as none of the mainline maps are
anywhere near large enough to trigger this minor glitch.


Former-commit-id: 4cd74057ad35cddaa79e7103b3af8bbb8cefa82c