Map screenshot feature for editor and game (with units):
The render engine is rerouted, so most objects are on it. So keep the
UI parts that you can hide by moving your mouse (footsteps,
highlighted hex etc...), may be useful. But don't write map labels
(not hideable).

Need a warning because filesize can be really huge (bmp). Dezoom helps that.

PS: a nice useful bug show the minimap rectangle change to include all
the map during the screenshot

Former-commit-id: 699d45d6d5a40aa4e1075da27f6ff34680f9ef87