race should not depend on game_state.
This refactoring tries to fix that. It creates an own class for the
rng part of game_state that is used within race. This still needs to
be checked (Mordante will do that soon). If this commit causes all
kinds of weird problems (i don't think so, but you never know), feel
free to revert it.


Former-commit-id: c71b97116c4ae088dc8b005390a5c589e7ba9372