Fix a little UI flaw spotted by Irrevenant in the forum:
Try to quit a new map with unsaved changes, click "yes I want to
save", then cancel the save dialog (to check or change something
before save/quit)

This causes an exit with the loss of your last changes.  Now the
cancel abort the whole exit operation. This is a rare case, so better
ask to quit again than lose work. Also seems to be how other editor
works.


Former-commit-id: ab89d024c2fac9f110eb715a5103d65a1c9e84b0