Fix an old 1.2 bug.
At the end of a move+attack, shroud is cleared on map but the minimap
is not refreshed (need another move to trigger it)

There is also another old bug, we clear shroud only in the attacker
survive the attack. I think it need to be done directly when
confirming the attack.

But it's a more invasive change (sighted event in the middle of the
attack action), I will check it later.


Former-commit-id: 70569f187f6f7522f166a8f40d459269175559e0