Minor code clean up.
Added namespace 'rand_rng' to random.[ch]pp and changed callers
accordingly. Functions declared in random.[ch]pp are left in default
namespace.

This was part of a potential RNG bug hunt which turned up empty.  The
namespace helps prevent some potential ambiguities as we have classes
declared using lower case and methods of the same name and case. The
namespace allows for non-ambiguous reference with the class.


Former-commit-id: c1ce93d66680113c9a49526a4e88e4b580a0f2a6