This change removes lifting logic for pre-1.4 versions.
If you need it, check out a copy of wmllint from the 1.4 stable
branch and use that to lift before running this one.

I did this for a policy reason; I wanted to kill off the --oldversion
switch.  It will *not* be restored; in future, changes to WML syntax
*must* be forward compatible in such a way that tags from old versions
can be unambiguously recognized (this will save everybody heartburn).

As a virtuous side effect, this featurectomy cuts wmllint's code
complexity by over 50%, improves performance by about 33%, and
banishes some annoying behaviors related to the 1.2 map-conversion
code.


Former-commit-id: 4643c976024d8b33dca1bf07b0588b10d4951ef2