Greatly simplify quote scanning in WmlIterator.
The old code tended to break on multiline strings ending with '"\n',
and was too complicated for me to debug.  The new code behaves better
in general, though it will throw an error on one pathological case the
old case seemed to handle -- single-line string followed by a comment
with an unbalanced double quote in it.


Former-commit-id: b66c46542caaaf769c1d14c4662548669af9f591