Add a new debug class for the widgets.
This debug option can output dot files which can be transformed to
images so it's easier to get debug info about the widgets in a
window. This function is enabled by the --enable-debug-window-layout
configure switch. It has not been implemented in the other build
systems yet. (Not sure whether or not it's needed.)


Former-commit-id: b4137ebbf96a75704f60e754da09ae60db21c10d