Rename the Gerrick sprites so they reflect the unit-type WML's evolution...
...to a generic Infantry Commander/Assault Commander that might be
usable elsewhere -- Sir Gerrick keeps his portrait, though.  Possibly
it should be the generic portrait for this unit type, but we can worry
about that if and when this unit goes to core.

Also, actually use an unused defense sprite, so now the Infantry
Commander and Assault Commander have slightly different defense
animations.


Former-commit-id: 3a4d276f17e2860452e17fe2f5208528c548e899