* New image-path function, ~SCALE(W,H), which scales an image...
...down or up accordingly. Skipped, zero or negative parameters are
assumed to be the image's original dimension in that direction.

* Adapted the _main.cfg of EI, LoW and SotBE to take advantage of this,

* by scaling down their high-res splash graphics to 205x205.

Note that ~CROP() and ~FL() take precendence (in that order) over
~SCALE() to try to cover the generic use case. Ideally the image-path
mechanism should be changed to obey the WML coder's will and the order
of appearance from left to right. It's probably best to postpone this
for 1.7 because it may break some WML that works on the assumption that
the current inflexible behavior is not changed. Obviously the original
author of this system did not predict that geometric modifications would
be eventually exposed to WML.

I cannot blame them for that. :-)


Former-commit-id: bcf571f35adb3ef4354e5d8567546b6f4d7bef9c