Heavily rewrote the way image path functions are analyzed...
...when loading images. Now all of them are C++ functors, and, except
for ~RC()/~TC(), they are accumulable and applied in a left-to-right
order from WML/C++.

This seems to have little, if any, performance impact in -O0+debug,
-O2+profiling and -O3 builds for me compared to the former
implementation. A test with gprof reports in average 0.0 seconds spent
in image::load_image_sub_file() for both implementations unless I'm
misinterpreting the data.

Tested in an AMD Athlon X2 Dual-core QL-62 on Debian lenny, non-stock
kernel 2.6.28.1 (SMP, model optimizations) running on native 64-bit mode
with the first scenario of HttT and a lot of hyperactive gryphons and
undead. 2 GB of RAM, 3.74 GB of swap. The compiler is GNU g++ 4.3.2.


Former-commit-id: 6e962a73f871d4f695d6e8a45cc9ab3021481931