Two more fixes discovered while checking translations:
"Edition" -> "Editing": The former was used incorrectly instead of
the active verb.

"the godly city of Tath" -> "the goodly city of Tath": some non-native
speaker probably thought "goodly" was a typo.  It isn't, just an
archaic form of the adjective "good".

In both cases it is not really necessary to change the actual
translations, but the string IDs containing these should have the
spelling fixed as .po files come back in for integration.


Former-commit-id: 52883f34ca0792eb8e17f9eb5f49aca2243f65a8