This fixes a replay bug that is mainly caused by events not being triggered.
The fix replaces replay logic with actions.cpp::move_unit, which is
pretty intrusive. I have done some testing and it looks good, but if
we run into serious problems because of this we should consider
reverting it again.

There is also a good chance that this fixes bug #11105 ([endleven] not
working when triggered on a networked client in mp campaigns), since
this bug is most likely caused by missing moveto events in replays.


Former-commit-id: 5c2133a29729b9d339aafe947cf4ecef9d1ddd59