Remove music files from the playlist when it is proved that...
...they cannot be located in the binary paths.

This is a extremely important performance saver, specially for users
who don't install the wesnoth-music package.

TODO: make the music_track fields private and make a setter for the
'name' property that can handle this stuff and cache a resolved path
or invalidate the track so the thinker can ignore it and save some time.

Former-commit-id: 3ecbfd3d27a38082373e7903c4caafa39a07fca4