The ~RC() image functor does not accept the special palette switch...
...~RC(palette1=palette2) syntax anymore. ~PAL(palette1>palette2)
should be used instead.

No idea what I was drinking when I decided to use ~RC(palette1=palette2)
instead of introducing a new functor (most likely I was thinking of the
overhead of the string comparison, but it's been reduced a lot since
then).

A wmllint test is missing, as well as backwards compatibility code. I
assume that this feature hasn't been used at all anyway.


Former-commit-id: 98cf2e895d8e4b480879dbf9f7b64a145e40bc6c
Former-commit-id: 87e8d66e7e7b0295d5314435a62416ea4965bc87