Pango-enable message= attributes within [message].
This patch includes (a) switching on this feature in the C core (and
diabling grotty old Wesnoth markup in this context), (b) an upgrade to
wmllint to do most of these conversions automatically (e.g. in TRoW),
(c) hand-fixes for some unual cases (in the Tutorial), and (d) a test
of Pango <i> (in THOT).


Former-commit-id: 5e333f70744614da4b93f12417cddba34c7b5d12
Former-commit-id: 3c3a504498f86bc771ca116623834d5860fd3065