Savegame reorganization Step 1: a simpler interface to saving and loading.
Providing a new file savegame.cpp for dealing with savegames. This is
only a first throw, much of the savegame functionality will have to be
transferred to this file in subsequent steps.

Disclaimer: This is a pain to do without breaking anything. I'll do my
best to test everything before i commit. However, expect some stuff to
slip through now and then.

Former-commit-id: 074d2cfb96ee0fb2651bfb6d45b28447b52d199c
Former-commit-id: 839aadeb382bd1cd3de40efd2e06411ad27eac4e