Fix continuity issues created by Chantal becoming personally important...
...to Delfador - happens in Delfador's Memoirs, affects Heir to the Throne.

This patch bumps Chantal from Druid to Shyde, and makes her death in HttT a
loss condition - though one that is highly unlikely to trigger given the
usual flow of battle in the scenario and the hitpoints she has as an L3.

This patch also adds dialog to both Elves Beseiged and Siege of Elensefar
that improves contnuity with LoW, with DM, and internally in HttT itself.

Finally, Chantal's name becomes a global spelling exception.


Former-commit-id: 97993f3c1ab8a338722e576358c2f54174aa1415
Former-commit-id: 9a08417733803497c75d44833e339833a4c603dd