Terrain types can now have description attributes.
If present, these override the name attribute for the "outside"
portion of the human-readable terrain description, e.g.  the part not
in parentheses.

The main use for this is editor palette tooltips; they now say, for
example, "Summer Deciduous Forest (Forest)".  I've also wired things
so that the normal in-game terrain report uses description rather than
name when it's available, as a demonstration (that's the diff to
generate.report.cpp) but it might be that level of detail is not
actually desirable in the game itself.


Former-commit-id: 2a92b20fd80ad7e81f22e63e493e1b030c746a30
Former-commit-id: f77b9e797cc9d799f39aad70e1281a5a072630a6