Add a helper header for a small test.
Quoting the header:

This file is being used for a small experiment in which some private
functions of tgrid are no longer in tgrid but moved in a friend class with
static functions. The goal is to have less header recompilations, when
there's a need to add or remove a private function.
Also non-trivial functions like 'const foo& bar() const' and 'foo& bar()'
are wrapped in a template to avoid code duplication (for typing not for the
binary) to make maintenance easier.


Former-commit-id: f882358e0e40151c02912c40ccb93344f85fdd2f
Former-commit-id: c9d68f54aad6ea769e4f56c405bc1d3cc94de37d