Enable the experimental storyscreen code.
The --shadow-storyscreen switch is left for experimental features that
will follow soon (including a new idiom for end-of-campaign text
screens).

Note that the new storyscreen code causes a noticeable delay on startup
for HttT as it caches scaled down versions of the background graphics.
We should probably do loadscreen calls there...


Former-commit-id: fd84e677d73f78da644743e785d79db446837027
Former-commit-id: 74d821e32de4493893737920ec10f0e5828421cd