Add parameter for the fake screen.
In order to make unit testing with different screen sizes possible it's
required to allow the setting of these values. Decided against doing it
in the constructor since I want to test on different screen sizes so
will rerun the function to get a different size. (Unit test code will be
committed later.)


Former-commit-id: 3d37e445fb5d20e24c1a134132d007507d68f0a6
Former-commit-id: efbd9ff1726b28f2edab41f399428c6e1f57d0dc