Replace the layout algorithm.
The new algorithm seems to be at par with the old algorithm so make the
new one the algorithm used. The algorithm still needs some polishing but
the big problem is the old algorithm which doesn't know about the new
scrollbar mode. So once this algorithm gets a bit more testing in trunk
the old algorithm can be removed and this one can get polished.


Former-commit-id: 7314d1dce3f5e21e6c514dd73c0be6787f4ace95
Former-commit-id: 066531795683eed5d41d0990ad19d1c1b99b8b81