Fix an old optimization of off-screen rectangular invalidations.
We already had a mechanism to avoid invalidating adjacent hexes during
the update of a big or not aligned off-screen sprite.

But a flaw with the handling of empty intersection with the screen
prevented it to fully work.  (and in some cases caused few incorrect
invalidations at the edge of the screen, which was very bad because it
caused blitting)


Former-commit-id: e8133901a2390ad1126f389afe59719080461704
Former-commit-id: 6e6eb8d473ad1562145c9a8af80d10783fc04904