Added a minimal implementation for the tooltips.
Since the tooltips are currently only used for the multiplayer menu the
have been impemented so they work there.
Post 1.8 it will be implemented better but it might be that the tooltips
then become separate floating windows. Since that's not sure added the
minimal version for 1.8 and look further after that version has been
shipped.


Former-commit-id: 5313f574e5efc3a48aa5b9ff7745c5a4d9d7d637
Former-commit-id: fcefa3a334ae4290eb06d720689cd0e1a5357221