Fix double click event dispatching.
Soliton ran into the problem that when using his touchpad to double
click on a campaign it might directly select the difficulty and dismiss
the next dialog. When using a mouse the problem didn't occur.

The output of events for the mouse were:
Push DOUBLE_CLICK_EVENT.
Process DOUBLE_CLICK_EVENT.
Window dtor.

The output of events for the touchpad were:
Push DOUBLE_CLICK_EVENT.
Window dtor.
Process DOUBLE_CLICK_EVENT.

This explains why the events are send to the next dialog. It doesn't
explain why it happens, but that might be the touchpad driver or OS.

The problem is solved by removing the extra events after closing the
dialog. Avoiding the events to be generated is tricky since a gui2
dialog can open a gui1 one, when that happens double clicks need to be
enabled temporary and then again disabled when that dialog closes.


Former-commit-id: 43dc0e43bd6f9dfeb00576402e2fd57041fb20df
Former-commit-id: 173278ef967f0b03df71072d0863078214b1c4c2