Add an overload for set_row_shown.
Hiding one row at a time in the game_load dialog, when the filter
resulted in an empty set was slow. Every time a row was hidden the next
row was selected and its callback called. Directly after that the newly
selected row was hidden as well... The speedup is quite noticable.


Former-commit-id: d1135e3e28ae684c1a97cafc2a97f23effa00f02
Former-commit-id: 4c51e60eb112b6dfd7cf3280dba75314bf6e48c8