Non-human null controllers are no longer set to AI.
This avoids dummy sides to get a turn, like the statues in 'Caves of the
Basilisk'.

Regarding compatibility: I did some minimal testing with a version with
and without this patch and it seemed to work properly. Since we're not
in compatibility freeze I didn't do more extensive testing. (This fix
also allows me to properly fix another bug [1], to be committed later).

[1] http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=563310


Former-commit-id: 9e3c1187c3005f3a8e026e2e1ca115ddc6b705db
Former-commit-id: e5ca16ad493534515357fe765b1ba61602cc60f2