Fixes some listbox assertion failures.
Ilor discovered them when showing and hiding items in the MP dialog.
This change also moves the sizing burden to the listbox, this was
documented in the contract but not done yet.

This allows the multiple item version of set_row_shown() to update the
size once instead of for every call to generator function.


Former-commit-id: ae8bd2589c5f8f661dd1b9737c1219ad2437d2fd
Former-commit-id: 78f7c83d56e74a55fa7c3abcfa3223ec468077d7