Add experimental option.
This defaults to 'yes' for odd-numbered (unstable) versions, 'no' for
even-numbered (unstable) versions.  It controls an EXPERIMENTAL
preprocessor symbol, defined only when experimental is 'yes', and an
EXPERIMENTAL automake conditional.

This is meant as a guard option for experimental, developer-only code
that should be disabled in stable versions.  The immediate application
is fendrin's new pathfinding/tunnel code, which we want to carry in
the main codebase for the remainder of 1.7 but be automatically
disabled in 1.8 betas and 1.8 itself.

Implemented for scons and autotools but not yet cmake.


Former-commit-id: b5f8ca90de995747581c7365ae4b89e0280afb7a
Former-commit-id: 336a011d7fec610de4984922c2f45bb9c6f3db79