Proof of concept to translate network errors.
This way network errors can be shown localized, after the string freeze
is lifted it should be possible to translate more strings. Strings with
variables can't be translated this way.

Changed the string in network.cpp to one which is already in the
translation database (so this commit shouldn't break the string freeze).
Tested with the German translation and the connection error is now shown
in German.


Former-commit-id: 02c8d19f81c5417b163c25368ddd46585aecf12e
Former-commit-id: c6ff550ce14bb7e02b09d167b387681610bf0515