Fix assertion failures in the tree view.
When opening or closing a node the sizes are now calculated better as
before and no longer cause assertions failures due to out of size.

Note the code still feels a bit hacky, but it seems to work properly so
rather commit now as rewriting it. The latter is scheduled to do later.


Former-commit-id: 36b53a790c5e42165e695283c9718ec9bf5f6a68
Former-commit-id: 7b8c6e8169d712dc1c85fb9d9ea502c0fc3d2f14