Fix a crash when using the scrollwheel.
A while ago I allowed the container area to be larger as the content.
This change allowed some bounds checking problems, allowing the scroll
outside the visible area when all items were visible. Now test whether
all items are visible, if so only allow scrolling to the beginning of
the list.

Fixes bug #15156.


Former-commit-id: 7e96d12bf9bf57a67f15788ac927813418de10d2
Former-commit-id: 69ec2a0fe4e9b39602360dd78f9944d1062a245b