Fix a drawing glitch in the title screen.
The --new-widgets version using 800x480 with the German translation had
a drawing glitch*, where the background image wiped out a part of the
buttons on the right. This glitch is fixed. Another one remains the logo
is drawn over the area instead of below it, which will be fixed later.

*) other resolutions and translations might have the same problem, but
this was the easiest to reproduce for me.


Former-commit-id: 01532ffe0aae841331c5c26be40a6d782b15739d
Former-commit-id: 50f3e8cdb0e76a7fce9d11ba87745df5c086764b