Now the top item in the stacked widget is selected.
Also change the order of the layers in the new title screen to make sure
the proper layer is still selected. The wanted side effect is that the
logo is now drawn at the bottom.

This changes the behaviour or widgets with multiple selections, but the
stacked widget is the only one, so no problems in other places.


Former-commit-id: 5c78f408129bb742f6780029fddf7b42ae99948a
Former-commit-id: a3250b9221a21d59474e57ae8c5b2021036a4547