Little bonus feature to check out terrain defense...
...for the units you select when it's not your turn.

Feedback needed, if people like it I'll do a bit of code cleanup,
otherwise this revision will be easy to revert.

This feature will be useful for the whiteboard's out-of-turn planning,
but I can always re-implement it specifically for the whiteboard move
arrows.

Note: the feature doesn't work while enemy units (at least AI units)
are moving/attacking.

Former-commit-id: 1961fca367a64b3ae10eca023950b304c1fa5ea3
Former-commit-id: 88a71d7fc095373b635dad45279941008d424390
Former-commit-id: fc68ea9c741ecd583b8959e68f92b5d0cb9ed6a9