Updated ana's implementation of get_bandwidth_stats()...
...to issue the string corresponding to network stats of the last hour
only. This is the original intention, as is visible in the
implementation in network.cpp, lines 961 through 969. Also added a
parameter to ana_components' method get_stats to select the time
period for which stats are requested and performed an analogous change
in the network_manager's method by the same name.


Former-commit-id: c10d8b87fcd3f5690a9e053546e877c6cb72a1ab
Former-commit-id: eeab0d9f5c69a26fdfc24b270b1b86a0516df50d
Former-commit-id: e9a82711f1d9b36ac8bb46d8c332bd218bcba855