Stopped the Join button in the lobby from always being disabled.
The problem was that the server was sending the open game slots as a
string of the form "#/#", which lobby_data.cpp then tried to convert
to an int with: game["slots"].to_int()). The result of that was
obviously always zero.

I made the "slots" string to be written as only the number of open
slots, instead of using the form 'X slots'/'Y total'. Change it if you
like, but now everything seems to work as intended.


Former-commit-id: 59e6a71fd13cd4af8491cb6f0b161e034a119583
Former-commit-id: 049e9c43ef833895da113cf6d91c33fb81a2ab08
Former-commit-id: e1d878da4720fe87168160adb6e93f2d14f9e510