Reverted part of 2011-01-08T20:40:28Z!koraq@xs4all.nl, since it opened a major security breach...
...in Wesnoth.

Indeed, it caused the Lua engine to skip unwinding of its state in
presence of an exception, which opened the way to several attacks.

For instance, setting gc finalizers and then forcing GUI2 to throw an
exception would allow a multiplayer scenario to execute arbitrary code
on a remote client.


Former-commit-id: 2f45bef2cd9269f568c48539c374bcea832fc229
Former-commit-id: 638be0009eb27079ababfc6d0eb28c99f24e4d23
Former-commit-id: e8e413ae1dce134577e5d5a41a7d349c1f2a5336
Former-commit-id: 8592a660e8d29348c336014b0fa8820cf09b9b5b