Reimplemented the tooltip event code.
The code now lets the widget show its own tooltip instead of letting the
distributor figure it all out.

The commit also adds a new event type class; the message. A message goes
from a widget to the window (instead of the other way around). The message
class is used to let the widget ask the window to really show the tooltip.
The reason for doing so is to both decouple who needs to show the tooltip
(no reason why the window is responsible) and to add the new message class.
(I've some usage ideas for the class, but nothing to commit yet.)


Former-commit-id: 21115b5964c172175e27e63f586f15371358d112
Former-commit-id: dc22d9fcfbac81898c65a946d5de9e228a741d82
Former-commit-id: e12edb9a2d8194d5fb82d1edb4f3f6899f633836
Former-commit-id: 83e50c7dd13d57f75d4222219e39f3bf645e65b5