Fix unit_defense report generator doing unchecked map access (bug #17721)
This commit fixes report_unit_defense() to return an empty report
object if it detects that the game_display::displayed_unit_hex()
return value is not a valid map location. This problem is not really
recent, but fendrin's odd usage of the engine to display unit type
info on his recruit dialog unveiled this flaw in the worst manner
possible.


Former-commit-id: 29b9b377b06760c15d06046b8563348d96ba5490
Former-commit-id: 67ef0a23da0ebc63f1fae528e8b82f2ded08f43a
Former-commit-id: c2b011e3f999b6a16e664b5027aaf805c910aa0d
Former-commit-id: 46f1d033f2c2e53f0beccf842db6daa83c109269