Added a new script: data/tools/terrain2wiki.py
Same as data/tools/terrain2wiki.rb except it works with current format
It also works with arguments, so you can choose source and destination
files

Still need some changes but it already works. Only feature that was in
ruby tool and not in this one is stats from. It will be added in a
future version.

mvt_alias and def_alias will also be added in a future version in the
alias column.

Another thing to not is that the tool does not parse terrains with an
'hidden=yes' because these shouldn't appear in a terrain reference.

Pep8 and Pylint results are good.


Former-commit-id: f87807a4b686b1d92c10e95000db7bbf669acf8e
Former-commit-id: a903fdfbbce593d87bd502369b79130c502aa0da
Former-commit-id: f577afb0cae88ca3ace3b8ccef00bec0fcdb1dac
Former-commit-id: 9bd66b968a9dd5dcf1958b6282605c59171e2d9b