config: Treat VERSION specially when probing cache invalidation scenarios
It's not explained in the code at all, but I guess the point of
invalidating the cache when a non-empty or non-argless macro has been
defined is to make sure the cache was really automatically generated and
not tampered with by the user. In such case it should be safe to make an
exception for VERSION, the first engine-generated non-empty macro so
far.

I can't imagine a situation where this would break any other assumptions
at the moment, so I'll leave it to the testers.


Former-commit-id: 6523ad834ddaf7ed5e532f554fecd5e53d7d7c74
Former-commit-id: 9462d48655b4ad4a7bbf7e6ff7d3559bb86d40d4
Former-commit-id: 4736cbe8067e8d97837817d1b0bfffd2acbfb063
Former-commit-id: a50f3a54114e1253fe15cb77d5a1fd45462657c1