WML preprocessor: implement #ifver, #ifnver MACRO <op> <version>
Currently, the specified MACRO is not expanded further, so if it
contains code inclusions with {} those will not be resolved.
Furthermore, MACRO should not take arguments. The whole idea is that it
should be defined to simple text such as "1.7.5" like Wesnoth does for
WESNOTH_VERSION since 2011-03-22T06:40:15Z!shadowm@wesnoth.org.


Former-commit-id: 3d3d0b126e17ebe98636fca5802a6acf0187394a
Former-commit-id: f9051b5b24d6129bc5100e4f32b63f8714dd020e
Former-commit-id: e3801dc406900078cf75e46f0d2fb76951329aac
Former-commit-id: 4b0b5c55f6d8f3e44df285afb95a735d0d3e721a