Ported the [allow_recruit] tag to lua and removed the c++ implementation.
Create an accessor for the help systems known units list for the lua
coder.  (used by allow_recruit)

Added an accessor for the new unit attribute "extra_recruit" for the lua coder.

Implemented [allow_extra_recruit], [dissallow_extra_recruit] and
[set_extra_recruit] by lua defined wml tags.

AI awareness is still to be done, the code locations are marked TODO
in unit.cpp.


Former-commit-id: 02f1a83557703b43f38bae27570885afd1dd3274
Former-commit-id: 58f36e775712d337af68c451bbab08e911f5d0d3
Former-commit-id: 0be006318c49f94fa4ab6514a10a3c5b19882f6c
Former-commit-id: 0391afca614fd705654ff5cf04fc340b52a983ba