removed Dev-C++ project files
Nobody has been maintaining them since more than 2 years.

It'll be a ton of work to bring them up to date (so it's probably a
better choice to convert the CodeBlocks files instead). The Dev-C++
IDE itsself is a dead project since 2005. Nobody will think there's
need to update these files anymore when adding source files. The
compiler used was MinGw, so CodeBlocks is basically a replacement.


Former-commit-id: 1d21b12ef4007f7ee0560d0fb0fce0663aa9b561
Former-commit-id: 4ebc16a6a0ecb8fad74ec0fe580603bb3e1c360d
Former-commit-id: 4da5a63933f1af2a5877387925c57a5055088726
Former-commit-id: 0c8af0ca3576efc149ca6649d97d0def27c3a2f3