Re-rationization of terrain names, part 1
Eleazar's "rationalization" of terrain names in 2010-10-22T20:51:37Z!me@jwbjerk.com,
2010-10-22T21:27:44Z!me@jwbjerk.com left quite a bunch of terrains with editor_name
attributes but no name attribute. While it may theoretically make
sense to inherit the name attribute from an aliased terrain, this
creates new problems of its own when a terrain is aliased to more than
one type (which one to choose for name inheritance?). In my opinion it
is best to solve this in WML and not add more layers (heh) of
convolutedness to the terrain engine.

Former-commit-id: 4f0b8906a5d1a125b22fe283c6288107bdbfcb62
Former-commit-id: ceade472fb7a6af205ce3fc45da1c2cc401f5cb6
Former-commit-id: 7113346d4bd7d973aba7f4fcd898f34e80f3ced3
Former-commit-id: deb91b8f7bc55026d548c5e158c62a3bb49a8de0