Fix network communication between wesnoth and wesnothd, bug #18632
This is not a final fix.

wesnothd expects attributes to be sent in alphabetical order (sorted
by client) to reduce server load.  The new config attribute table is
no longer ordered.  The interim solution is to sort the attributes
before sending them. So multiplater client bear a sorted cost no paid
by single player clients.

A much better solution is to fix the server to not require sorting at
either end reducing network and CPU load.


Former-commit-id: 60dbf53398c6e09734b35d521edfa17ab3af8262
Former-commit-id: 07f4e13d83975cb4b869cc6fd5b30b6f07950f67
Former-commit-id: 11b22a08b0c93720c8ab7fc3823255c9f3105c1b
Former-commit-id: da7ef58e16454d0165ff1f5a19f1024e4ad995b1 [formerly 95817b411f579b3c0f1424af010b70834b7fe2d3]
Former-commit-id: 1f907f3b6b833adf9dfc76a0f4d0e0e928efe9cf