Fix implicit static cast...
...of config::attribute_value & to tstring & or t_token reference
causes crashes.  Some compilers were generating a temporary object and
binding a reference to the temporary when creating an implicit cast to
an attribute value.  I removed the implicit cast and replaced all (I
think) instances of the implicit cast with explicit 2 step operation.
This fixes bugs like bug #18663, bug #18684


Former-commit-id: 6a51301db961ae557d2001c4b39c666524c90a24
Former-commit-id: e29e9732aa760f669150ca8b67d1c26d2b67cf4d
Former-commit-id: a01ddc760bd419ffae543e1f632b3429c3581933
Former-commit-id: 07103f639fa4125290162a911053b293d327ed0f [formerly 0d199daf717949b55cbdc690cc988fa10a270d22]
Former-commit-id: cef4c1969b8b21b4db792abdae383184e8eceda7