refactored when (and when not) buffering of added events is done
(fix for bug #18695 and bug #18737)

The reason for the bug was that buffering was activated and deacticated
at wrong times.

Events added (or removed) during the current event call stack do not
have effect until the base event has run out. The code previously to
adding the [event]id= feature behaved this way and it is now (again)
this way. Although it's still not exactly clear to me *why* this
buffering is done, adding event handlers which were just added during
the current call stack already to the active ones (FR bug #18713)
would require modifying the array of active handlers withint the same
loop where we are iterating over it, which is always a problematic
thing, and this in an area where every mistake quickly results in
fatal WML engine bugs...

This should be subject to further testing with scenarios featuring
complicated event structure envolving custom events, [fire_event]s,
event ids and removing of events.

The reason for me not being able to reproduce with my testcase was
that I didn't have shroud enabled. ;)

The reason for me being able to reproduce only sometimes was (I think) that
originally, buffering was initialized with random data so it sometimes
delayed the preload event and sometimes it worked correctly.


Former-commit-id: cc81aa2193045552f44d445a5da523214b4d881a
Former-commit-id: cf7ba0e5f26fc1e412b3a363dee572a38c2f0ccc
Former-commit-id: 0882590273a919f61c1634ce73b5cbd6948dd2c9
Former-commit-id: fc4b77a7a7ca97ca9306489b0ffe8f49696d1b59 [formerly 70c27316ad6df87191fa1ea190017f478a2227a4]
Former-commit-id: f2ba10875a1ee29f7dc755d33afe1a54d38e95bb