Fixes #bug 18715
(Game hangs at loading screen after flushing caches and save index and
pressing F5)

Previously, when the .cache/wesnoth directory was deleted, while
wesnoth was running and then the cache was refreshed with F5 from the
startup screen, wesnoth hung.  The crash occured because wesnoth
didn't recreate the cache directory after startup, when it is
recovering from the cache not existing.  Now it makes a new cache
directory as needed.


Former-commit-id: edd7ee08ad703dd8578bf844f48045aa8a3527c3
Former-commit-id: 69d4f320120987e40ef48bc884b8867993f6baca
Former-commit-id: 1799da315faef0b3cc4856d4a474b82b9a43d8dd
Former-commit-id: 849f0179c3f285cd854a3e64e88053925dff9426 [formerly 4a91b7601d207ecb2471e36aed017026034e8b41]
Former-commit-id: fc90382d487cd52793deb360d58ba1af9b28e814