Fix for bug #18745:
(It is now impossible to jump to ability topics from the sidebar in
most translations)

Changed all vectors of ability_tooltip objects to be vectors of
t_strings.  Previously, sometimes they passed std:string and sometimes
t_strings. This meant that the help index was a t_string->translated
string ->t_string ->base_str(), where the base string was actually a
translated string.  The tooltip was passing
t_string->translated_string as index which, although inconsitent
worked.

Now when the tooltip is created it uses the translated string as the
visible string and the base_str() as the index into the help, as
intended but not done in the original code.


Former-commit-id: c76e7f61f6de04cf4eb808d67e26f6328106adf7
Former-commit-id: 82e60b521e864625f413ba6c03c56f5354e2512d
Former-commit-id: d1e8429f63c4a66494cd6b0a2666d1c0406b406d
Former-commit-id: bf64ef829efc4dfed2bfc7a889e23a314286ec1c [formerly 71888e9921d45ac3ffc4b90f53387a3a73c80a25]
Former-commit-id: 2266b75b996690bec6b7cec29dbd1adda6fd4ad7