Add a work-round for bug #18534.
The destructor of the class restored the moment of the units. When a
unit no longer exists it overwrites random memory. The issue has been
found by valgrind.

anonymissimus tested the work-around since he was able to reproduce the
issue easily, while on my system it went well for a long time.


Former-commit-id: 691e02358b7bb257cb5c3b9b289e0a805d22c61b
Former-commit-id: eee17dc0b2634cfb307cfa594d0213ac8cf78e71
Former-commit-id: 4cad00c75e24a99fb8d423d5798359b93421ef2f
Former-commit-id: cd5eba635f83da2b473a35b76b5beb40308add8b [formerly 2557bc3c518ebc596c7800c64ff285107e53cb48]
Former-commit-id: fcf39524c2d381665a7b453352f1da4487755e60