Apply "Setting to ignore encountered_units" (patch #2940 by negusnyul)
The preferences entry name and tooltip are not the same as the patch in
the tracker, as we discussed on IRC that this doesn't exclusively affect
mainline units nor does it need to. I also changed the tooltip wording
to fit as much as possible in my 1280x800 screen without changing the
font size (works on 800x480 with a smaller font size).


Former-commit-id: 7240f8438fd39e0506344f197848497f2d035273
Former-commit-id: c6c60287db450c206b1c7237ca928a9d7d4aa6f3
Former-commit-id: c12310f3c3d12accc8719c5f7f358c9a23e13498
Former-commit-id: 0c9d0231a286440aff9ad38cdf31d708e60be409 [formerly fba7487807a1e7d0b88460f0fb35b5ec399a5275]
Former-commit-id: 6ab3be2494dc61a06a374b3470a7a4aaa9fc7864