set playsingle_controller::end_turn_ to false in linger mode
(fix for bug #16821)

The problem was that this variable gets usually set to false during
the start of the side turns. If check_victory() ends the level between
the side turns, this initialization is missed so linger mode thinks it
has already been ended.

Other call cases of linger() should call the function with
end_turn_== false already. That is, I think they already do, since if
not these would have fallen into this bug category as well.


Former-commit-id: 533448c8e821254ee7b5e6a40d423dbbbe008c0a
Former-commit-id: 9f61aceb4e17dfe8e7854173a5650617bbd32683
Former-commit-id: 517d28a8b5af951d9e1a0c173826d0ddd6561fd8
Former-commit-id: 79990f43dfafa3bf2a74e644dd79bfba1e6e02ea [formerly 3fb9d53f3be4cf1bc2ba2edde91af149ad873083]
Former-commit-id: 8eeb22e0d0841c689887af0bea6e8df6b573c7f6