Fix compilation with clang 2.9 / boost 1.47.
This applies a slightly modified version of the patch linked in
bug #18399. The changes are coding-style related.

The patch was intended to fix compilation with boost 1.47, but that
boost version is not yet in Debian so didn't test that part. Since it
does fix compilation with clang 2.9 Iapplied it, without waiting for
boost 1.47 to hit Debian.


Former-commit-id: 4063c5354e26f288e9d010702a041abc1c2d286a
Former-commit-id: 675e7feebe4f341b1d379a3a8408b1a059e57dae
Former-commit-id: b28da0618cce34ce4d5f2d1f1e827c7ec781e090
Former-commit-id: c9939964d0c5f2b54a212427df5e43d765131fb9 [formerly 1b0d694d1ce65888ecd1566e168df6e774f52442]
Former-commit-id: e1d94be574a5f9cb528545e5c1a82466f18c8cf4