Fixes crash when using fullscreen on Mac OS.
Fixes a problem when using the SDL nightly builds (post 1.2.14).

The direct blurring of the screen fails on OS X, since the surface is
not a neutral surface, so only testing for 32 bpp isn't enough. Exports
the neutral surface test from sdl_utils to have a proper test in the
canvas code.

This commit helps fixing bug #18319, but the speed of rendering is still
too slow.

Thanks to Alarantalara for testing the code on OS X.


Former-commit-id: bbbf0d38b552c477102e2e911b04490e111d836b
Former-commit-id: 9f577070f063bf0edf4259dc8d49139717c1626d
Former-commit-id: 3193837d95a7c5ddb174b7c745a6c179a38f97cb
Former-commit-id: 214670a2ac82568f6c3c0da49c250d5d7d25a8d2 [formerly a2691c9269e697699dbe9ba2d465565e505cb58f]
Former-commit-id: a7baae4147802c008f608108972483c855e29a40