Fixed bug #16571:
(Multiplayer login error with certain username or password formats)

The new username/password serialization format is as follows: @ <field
value> @ Old usernames/passwordswill be converted as necessary when
the user changes their credentials in-game. The user will to hit
#16571 once if they are using an affected username or password saved
by a previous version.

Unfortunately, the underlying cause is still around and this only
fixes this particular case by forcing new usernames and passwords to
be stored in a format that can only be interpreted as a string by the
config class. Dealing with it isn't feasible at this point, so I'll
file a separate bug for the general case and mark it as Postponed
later.


Former-commit-id: 110535dabfbdf496a90b6603b656df5ed58d0829
Former-commit-id: b43ca3bc0f86b6a2c9e32a80b68722a2552daa70
Former-commit-id: 576d092e5cd35d5f98c0dc332f0ff8ed8ce3cd8d
Former-commit-id: d04ba4469cda850c8d8483260f9fc2849afca8b0 [formerly f792a13690a0cab90f2ba620be7e80bdeb2d0143]
Former-commit-id: d3ddb8541d7067d113d84456abd9660edea3acd7