Revert revisions of mine: 2011-11-09T00:22:25Z!anonymissimus@arcor.de, 2011-11-10T21:18:53Z!anonymissimus@arcor.de
These were the variable substitution delaying in the [object] code. It is
no longer needed due to the workaround [object]delayed_...=yes|no.

Removes repeated WML parsing. Reverts behavior to mostly what was in 1.8
If ABILITY_TELEPORT is added by an [object], or [object][effect][filter]
contains $this_unit, delayed_...=yes is required for the modifications
to work as expected.

However, if the same object wants to get variables substituted and
add ABILITY_TELEPORT to a unit, the code would have worked in 1.8
but doesn't in 1.9.
(bug #18893, bug #19225)


Former-commit-id: a3d28a3d2676f63b46b177ca288ede7458c9fdf5
Former-commit-id: d148fa1347e53bc39ae413238194804ec4765353
Former-commit-id: dd959152c77727627a365f86b448b29b907007d9
Former-commit-id: 20f6672e6eb556e745c549276534d4c4734533c4 [formerly 3061143ec69d429f9f4fd53938a79487b0d911f5]
Former-commit-id: c11c13202b9e44f819cca5bcc69a6e6d711b3893