accept only non-empty variable names in [set_variable(s)]...
...and wesnoth.set_variable

This supposedly common problem (value of a set variable composed
of the values of other variables which perhaps don't exist or are empty)
causes corrupted savegames and makes :inspect fail. Hard to track on
wml side; there are probably a more ways of corrupting saves via empty
key or tag names.


Former-commit-id: 1b4224edc93c9a8c7676f1e3629ba05333e79199
Former-commit-id: 1c94c7035c51015981241dff7fbf162e78468eb2
Former-commit-id: 9861c5b2a26a3fcd2bf724b266e077c04532ab0d
Former-commit-id: c902e78bce83e127a3e57a8be5c4970e5c1b1cad [formerly 1ed5323ed12a1eef9481e2fe95506733bd6a3582]
Former-commit-id: 806f244fce893412f29f955303219f3358f145f0