implement better faction detection (bug #19281)
Whenever default era is used the units should be placed for sure now,

this was IMHO not guaranteed since the recruits are internally a std::set,

so the first entry can change.

In case that one or both of the players have another faction than the
ones in the default era, none of the players gets the extra units.
Support for Khalifate can and should IMHO be added (given a suggested
set of units + locations).


Former-commit-id: e8215e21a8e14dae9b036907cbd9508541e3ebf2
Former-commit-id: 0303e88e21a4f5e2319d9f272ee486ac2127eac8
Former-commit-id: beab5c3d8b8fa9051bbc1c14051dbe4025243bc4
Former-commit-id: ba20f87d6b44221d0f8a0b64807bd147739f889d [formerly ae90b0b163a33af03f296624debc304bdf2e1f0f]
Former-commit-id: 90f5108bff920ff03a15794e314c7146dde5f920