Added unit sets for the "not a default faction" case composed by Doc Paterson.
If any faction is not one of the default factions (e.g. Khalifate)
it gets these units now instead of both factions getting no units
at all. If the other faction is a default one it still gets the set
composed for it in particular in the same game.

(bug #19281)

http://forums.wesnoth.org/viewtopic.php?p=518542&sid=73489ef47cba8b3a39ac219353f64613#p518542

Former-commit-id: 1ba1e13ac96b45d21875ff68376d496c4c03a926
Former-commit-id: bb8ce42e0150961cd701b197d273f0018878db85
Former-commit-id: d4d5687094f6309747777736e2840f9231182cd6
Former-commit-id: 52e413b4fa3386c40c153ad98d6d49c229b92c0d [formerly b22c9505a453c323206f979a1887e9ea60eaab3a]
Former-commit-id: 2d4472f54c36e824813e19b20220ab7ffee922df