Fixed the translation of some tutorial messages.
The second part of the narrator messages were not translated, not sure
why. This patch is a work-around not a proper fix. Once 1.10 is in its
own branch a proper fix can be committed. That fix will change
translatable strings, and is the reason for a work-around in the trunk.

Fixes bug #19316.


Former-commit-id: 86ccdbffeeb077485677ba370a63acbf0b93e464
Former-commit-id: 1b955fc11526602b7f5b66d9c258a3d794241e1c
Former-commit-id: 81615f2f17e9c8e77c3dc499eb2394c1274ea791
Former-commit-id: 6b67139562a77b884964a2cc2b741be90b4d3103 [formerly b3f4987bca6fe0586cff2265cc611955c11a669b]
Former-commit-id: 418df0f129d6c6679cd4177f9544cd81c4ca4066