Restore leader unit image in Load Game dialog (bug #18683)
(Candidate for 1.10 branch.)

This functionality is back with a slight change to 2010-11-29T19:46:11Z!shadowm@wesnoth.org in order to
left-align the map snapshot only when the leader image cannot be loaded.
We also no longer spam stderr with errors about missing leader images
(which can be caused by unloaded campaigns or eras).

Anonymissimus pointed out this functionality was still found in the GUI2
version of the dialog (--new-widgets). Whoever last changed the saved
games summary code probably only bothered to update this experimental
code without fixing the normal GUI1 dialog. As I pointed out in the
tracker, the layout code was still in place.

As a compromise for the time being, we always assume the "magenta"
palette for TC recoloring now. This doesn't seem to be a problem since
the old code already made an annoying assumption of side 1 using the 1
(red) color range instead of permiting the use of the side.color value,
or reflecting the currently playing side number at the time of the save
generation.

I may later extend the saved games summary to include additional
information to get rid of the magenta and side 1 limitations.


Former-commit-id: 8d3bb25811a3c9c6650fb4f494e20cba9c7579ec
Former-commit-id: 2dba1d65bdc67b25bbecb1d987d1dd8cd7fcf9f3
Former-commit-id: 242910c1364ebe59a85a8c9da4ef681269b6f2fa
Former-commit-id: fde8331664276a176a60a0a40c59320f096a66d9 [formerly 4418e4eea5a6731c5cdc0ce76c2e77fb20b139e7]
Former-commit-id: 473cec62616849aeca0fa218ebce71c8c98bbb88