Remove a backwards compat hack left over from previous commit:
2010-07-18T15:21:41Z!guillaume.melquiond@gmail.com

The hack causes problems now, since I want to load helper.lua previously
already and the comment says that it should be removed in 1.11.

I *think* the reason for the hack was to avoid loading wml-tags.lua twice
in case that a replay and/or (perhaps) savegame is loaded which was
created with a BfW version previously to 2010-07-18T15:21:41Z!guillaume.melquiond@gmail.com, 
since those replays/saves seemed to have recorded the commands for
loading the core lua files, which is no longer the case since back
then. Those replays/saves are no longer relevant, so everything hints
that I can safely remove this.


Former-commit-id: c9b565c8a2caad7f0b225caa42658bbfca58f421
Former-commit-id: 13fa3df3f184dbca7a25b43e56ff494b9f2b4313
Former-commit-id: b780696d57fa087cd1b09dd34d83c892d7c4df39
Former-commit-id: 8dfcef1dd2002caa05bbbc5abb7041716ea32539 [formerly 06129f3998fe20e341d235e81ae2335337c8eec0]
Former-commit-id: 414d5a37ae64c9c94d5d632ffc0923a3435e32c5