Adds new floating point emulation code.
This code can use several fixed point formats or a double to emulate
floating point code. The code is still work in progress. The intention
is to replace the current fixed point code and use it in more
areas. Hopefully this makes Wesnoth run snappier on devices with slow
floating point hardware, read the Pandora.

By default it uses doubles, so when used in existing code there should
be no differences, between the existing code and the new code.


Former-commit-id: 282c702c551b7a13d32ff502c98381f82851ebbd
Former-commit-id: 56411d117a4478ccf02afc99742f0728040f07b8
Former-commit-id: b768bbe8f1d965bdca5acb8034b65328bd1e44dc
Former-commit-id: e465d0216d782aa5d9b40315cdde16687f5ddfa1 [formerly 1c204880666431a9ba050d175968f21b3249bcbd]
Former-commit-id: 506bbdbf585560a613135d1c6dc299700fbc2c04