addon/mg: Better fix for builds with non-GNU C++ STL than 2012-02-24T17:43:46Z!anonymissimus@arcor.de
The point of the at() member call was to throw an exception if the
access was invalid. By using find() without checking, we could end up
dereferencing the end iterator of the addons_list. Thus, a helper
function has been added now to use an assertion to protect against that
hypothetical situation early enough.


Former-commit-id: 41c8e6a75b6d46c2e36273116d17fa6401af557f
Former-commit-id: 58c91cb7530d299bff8a76038793a5977e5ce21e
Former-commit-id: 26309c6be33956b4c8bef7031b4f90f06886100c
Former-commit-id: d0293ffdc84bbb203fdaf9a282293bec78612863 [formerly f58ccf9b0b2f71fb68f4c8d5a1453b2ebe01f6da]
Former-commit-id: e8f1ee36c2bcd5f8b776516951ca9cd63d7d2c16