introduce inline SSF support in [objectives]
No deprecation needed since it defaulted to the 'default
objectives' already. They are used for a side in case that that side
doesn't have side specific objctives set. Use a SSF which matches
no sides or all sides to set the default objectives, if a subset
of all sides match, the side specific objectives for these are set.


Former-commit-id: d32b6a1a1f29a8e816bd05d96bfaabd79af1c15a
Former-commit-id: b310fdbc9ceccd5ca796fa9401d9822a110fb0aa
Former-commit-id: 678e0fe0212c87c8bc6cec0784024a6176707acb
Former-commit-id: aa332ba2da21db53c661a72033ce792321cf7a06 [formerly 80c8bf0e386fa4bd7bffe696dcce41c0df791889]
Former-commit-id: 20573fb60bdb4794ea6ae25bc6d2a31855ca6e0d