Improve the multiply instruction.
It now keeps more resolution and is less prone to overflows (unless of
course the result overflows). The code is aimed at an ARM v7 so might be
slow on platforms that do not support an instruction for:
32-bit signed * 32-bit signed = 64-bit signed

The alternative was to write code like for the division, but that would
be more code and slower on platforms with fast 64-bit multiplication
result.

The output of the titlescreen image now looks less dark as before this
modification.


Former-commit-id: e8260299b5340f4e896e2b1ebf7cb511e5bd0401
Former-commit-id: ef716e3ce3125476f7385caee742770a6bee0d8d
Former-commit-id: 5cef09fe4c54744de6b59eaaae22438018295d81
Former-commit-id: 4ea2281afebd1377a6e3a87f6b04cf5495129f77 [formerly fd310b34b0998676c40611b1d826a2e8e67a2ae0]
Former-commit-id: 6c477daeab9cb83ce41cb5ed5b698e9b47f8102b