Replace non-standard Underground ToD schedule picture (125x38)...
...with an upscaled version of the correct size (125x39), provided in
bug #19415

This solution seems sensible for reasons stated in the bug report. I
normally dislike upscaling graphics, but in this case the original is
crappy enough (it even appears to have JPEG artifacts?) and the
difference barely noticeable to the human eye.

Since this kind of changes can apparently break add-ons nowadays, this
probably should _not_ be backported to 1.10.


Former-commit-id: 6a02877fb9b693ca6cd9609ff279d98364ad552a
Former-commit-id: 19bd1e9ca37799060fd04cde209ce58f29cd333c
Former-commit-id: 93e3eb570e976f3e59c0e31c2ff9be5da520c392
Former-commit-id: 149c04dffd79e7acee904b87bedc7a77f6e5ad9b [formerly 697f4310dd47d93aacc598040c617327ef25f408]
Former-commit-id: a9f06f297202e5e13146444942d1e26d6fc3a113