addon/mg: Force version_info cache refresh after installing add-ons
The cache obviously becomes goes stale after installing/upgrading any
amount of add-ons. It's normally refreshed in the same steps as all
other caches (WML, graphics, etc.), but when we are in the manager UI
loop that code doesn't get a chance to run, so we must reset this
particular cache before continuing. This makes add-on statuses work
correctly.

In general, the version info cache was a terrible idea, but perhaps its
underpinnings could be adapted for add-on dependency/conflict resolution
in the future.


Former-commit-id: 8ccfa03d2274bf6973e9331021a1096aa7ebcb76
Former-commit-id: c57ab96b95801ba1692ece9b199fe2d3ded57fae
Former-commit-id: fb77bca839b23e322b5016281df1b07dcd178f74
Former-commit-id: aa2e17c099a7082212eccba9b82db05291c0f54e [formerly 2b2063f49feea4280b84e2b8e016f1be306ccdb6]
Former-commit-id: 0129691e13c6f3102232131f263d5580a05e005a