gui2/tgame_save_message, gui2/tgame_save_oos: Capture and grab textbox...
...values correctly, fixing bug #19562

The saved game filename textboxes in these dialogs appear to have been
mixed up with static labels in the C++ implementation of these dialogs,
so they weren't working as intended.

It was necessary to change some interfaces to accomodate this fix, but
it should be safe for backporting to 1.10 either way.


Former-commit-id: cfdc2f2cf725c5b8cb372e84941fd43a7f926415
Former-commit-id: 2478e9a80fe7074e6bafaa882ac36dfa115be655
Former-commit-id: 7d6c53f8f5672b309531029415bfc4c992a7dae7
Former-commit-id: 1d715aa563db4ea52aafef452b377bf462dc79e1 [formerly a666d9b5eb9ca052745e76a7f1bb658f1f34400f]
Former-commit-id: 269822eb36b5159c0266d6761cd789313231fde1