gui2/tgame_save_message, gui2/tgame_save_oos: Capture and grab textbox...
...values correctly, fixing bug #19562

(Backported from trunk, 2012-03-21T01:44:33Z!shadowm@wesnoth.org.)

The saved game filename textboxes in these dialogs appear to have been
mixed up with static labels in the C++ implementation of these dialogs,
so they weren't working as intended.

It was necessary to change some interfaces to accomodate this fix, but
it should be safe for backporting to 1.10 either way.


Former-commit-id: 47bbf2a1a6e61e9f7c22a36fec575a171a2985ab
Former-commit-id: c3ef6ca33508b7a521abcbd679cf55531c4d8a38
Former-commit-id: 21e5cbfa6ffd013af4e84926e8f35215735c0ad4
Former-commit-id: b5df7c498e3c07754cdd31d445acb25f7b6e75fe [formerly b63f6f7b5c85800ba057e4efcdc7d5ac1a97dcb1]
Former-commit-id: 52e92d269e8b7d77221f39a9818c6f613db08055