gui2/teditor_settings: Use a cheaper, less CPU-intensive solution...
...to refresh tiles after changing ToD settings

Instead of refreshing the entire display, this solution simply
invalidates all gamemap tiles and requests them to be redrawn, saving a
lot of CPU time and making continuous operations (i.e. dragging one of
the sliders around) smoother. We still need to redraw the dialog,
though, and there's a hack involved to avoid a floating slice of it
appearing in the middle of the process.


Former-commit-id: a00adfd85b13eadb664a9abae00d5d90b8ed9962
Former-commit-id: 4d7dc8cc643bc25d4a01dc5bd000164dd7d92099
Former-commit-id: 8b0116d3272f00afa0a062a0395ad8d8030153d7
Former-commit-id: c79cb31954adc9dc071043c817549fe3ac09b83a [formerly 405a96a8329859bdb7d980659e68921837a8ca1c]
Former-commit-id: 8338fc4ff7cd2df36be221db92036ac4228df950