addon/mg: Set the Add-ons Manager dialog title to "Add-ons Manager"
Previously this was "Get Add-ons" or "Update Add-ons" depending on which
view was active (the latter was used for the upgradable add-ons view).
This distinction doesn't make much sense now that the dialog is
constructed in the same path and supports a third view mode. Thus, it
makes sense to use a generic catch-all title now instead.


Former-commit-id: 4b190ff7593a17a5bd92395468215cbfe17bfdde
Former-commit-id: 3bd42592976026ff3937fc87df3bde48d8431e6e
Former-commit-id: b46c7863161fe008bcc6820fb95ac1bb088a210e
Former-commit-id: 06aca8ec1a7a134222b3f2100e40f61893715511 [formerly 66860407a6b60203cab5fcfc8978d2e36c9dac41]
Former-commit-id: 248a5150d49b310dc7b21eb4b239b6d9cc51d25a