set MSVC's intermediate output directories correctly
I finally managed to solve the linker warning spam about missing debug
information for liblua. Apparently the file MSVC was looking for was
overwritten by the one for wesnoth wich has the same name. This sets
an individual directory to write it to as it's already done for
wesnothlib. Also, this finally enables me to debug deep down into the
lua upstream source. :)

Also some other mostly warning related adjustments and fixes.


Former-commit-id: 68d94765208f54bf4063716cdb125f5d47558b55
Former-commit-id: b123b99afffb82afa7410723f509f7f24375360e
Former-commit-id: 7acf4c0a0baa924e2f3b3b6a6d47c93bcbd80b71
Former-commit-id: ade43ef2234f2cbe765f18092b2610045e5a7c7a [formerly 9e86e4b2cf611397a8e99298748a1e960708a67a]
Former-commit-id: 65b586dcfdaac1ac50946b5a04b1e008f501ab09