gui2/teditor_settings: Use a cheaper,
...less CPU-intensive solution to refresh tiles after changing ToD settings

(Backported from trunk, 2012-03-21T08:23:30Z!shadowm@wesnoth.org and 2012-03-21T08:29:52Z!shadowm@wesnoth.org.)

Instead of refreshing the entire display, this solution simply
invalidates all gamemap tiles and requests them to be redrawn, saving a
lot of CPU time and making continuous operations (i.e. dragging one of
the sliders around) smoother. We still need to redraw the dialog,
though, and there's a hack involved to avoid a floating slice of it
appearing in the middle of the process.


Former-commit-id: 5c6888eee3884204b5ff4277084eb24ee4e35040
Former-commit-id: 875df43e2cbfda18312336ee8c51ac572e494ec0
Former-commit-id: 9c5e38209ac4d68e3888711aff0627462ce97112
Former-commit-id: 563074cc8b347f05f9bfa1dbf533fcf68d8bab86 [formerly 195f648a19d31114d6318a564a7c6f183a1cde76]
Former-commit-id: fb2da6f7a5ea881aef28cee254a073b99c608860