fix $owner_side incorrectly set if capture event fired by unit movement
We are deleting the captured village from the set of owned villages of
the previous owner before querying which previous side owned it (and
giving it to the new side + firing the event), thus the result was
always "unowned".

fix for a bug reported in the forums:

http://forums.wesnoth.org/viewtopic.php?f=21&t=36569&p=525789#p525789

Backported from trunk 2012-04-07T16:43:03Z!anonymissimus@arcor.de.


Former-commit-id: a9060617201597035fbad3604f352b7269858d02
Former-commit-id: bc3b8f384e1c683297fd9e23dd4573ef92f3f2e1
Former-commit-id: dfaff17766561e96b6058e7c3f05c2d3bfa9ff19
Former-commit-id: 21ff1dbb23a09de0c683801758e36e67f0de2272 [formerly 56c57c2ba366aa77f24d44e6e3e1dff02d8baeb6]
Former-commit-id: 2434dcaad19146a65051891c996b984161ce3632