Fix binding to a temporary.
The result of std::string::c_str() of a temporary std::string was stored
in an object to be used later. Changed the member to a std::string to
solve the issue.

NOTE since the code involved contains serveral OS specific #ifdef's
not all changes are tested.

Issue found by cppcheck; it choked on the code.


Former-commit-id: aa93402fb0249ec3f52332a37723c46a05a4b464
Former-commit-id: 9cc6aba9a467ad7f729a10f4743a85b38fbcae7a
Former-commit-id: 7f3f39c3f10e5de7ac0afc291434f4ad60941403
Former-commit-id: da6e2899797d05b6e95dbc4d0b498f88c9747fe9 [formerly bdcfacad83ec538b7521095311fbe1c693b67ee5]
Former-commit-id: daf7056fd30a34b6be0d53fa8983b416ed9bb010