Make leader_image paths in the saved games index binary path-independent
(part of bug #18683)

This makes it possible to store in the save_index image paths that can
be used in the Load Game dialog even when displaying units whose
graphics are in binary paths not currently loaded. This alleviates the
effect of bug #19410 on the whole mechanism, but not completely.

When the save_index is missing, this won't work properly; we'll need to
add the leader_image attribute to the saved game config itself later.


Former-commit-id: 5f80d3b93941acb08665230ead57f32347530b88
Former-commit-id: 7776304ff23382a72c567e57c74ef4ea2fccc472
Former-commit-id: 9b6e961f5a4da816d0c547118495a5bdcb001773
Former-commit-id: 570c0d3af405e04273219f39e20e3e3122cc566f [formerly 8a8e52b7cf92758aab0038eff64f8f76ca8a974f]
Former-commit-id: 7c7b30a0ff03d6911caba1ea6cdd0625064625f5