Add the find_at functions for the pane.
This allows the usage of the mouse in the pane. Also increases the size
of the pane to have a bit more space to move the mouse around.

Note the offset to the mouse movement and what it affects on the screen
is intentionally.

The code is used to experiment with a different approach of the
implementation of a listbox.


Former-commit-id: f71cc8feb32040b5ff93633f10d1c6e0cdeec217
Former-commit-id: 770d62ea173342c588fcda81cbf09b306997dafc
Former-commit-id: e9368b27315f76a451a321f6c898dd23398e9bbb
Former-commit-id: 0bdf2b9ddafbd7e57783f1508ce78a4530ef3dca [formerly 64a4c5e17c47a0c49f3bba452c24fee20b38147b]
Former-commit-id: 3e692e038f9d0ffc581abfe45f17e88e36047147