Created basic implementation of the viewport.
The class can now hold a widget and forward calls to it. The class is
not yet complete but it looks much better now.

Also allows to switch between the naked pane and the pane in a viewport
with a #if 0.

The code is used to experiment with a different approach of the
implementation of a listbox.


Former-commit-id: 5633bee2bd1f24ddf8a3fe9c2b31f96e7dbec5fe
Former-commit-id: a634db4ef383e778b922a1d09d34b951d3a51367
Former-commit-id: 2fcd5e3b4bb50d4e3578453527ce19b9f90eeb7f
Former-commit-id: 240c464a81a03719d05cd726d756efb5491423d2 [formerly 7d76deeabc4942e95a956bfdc16f80a27f573f1e]
Former-commit-id: 1d1fb7ebbf3c065ead20fb70fb977675d0577f3c