Implement the expand and collapse button.
The items  in the adon list can now be expanded and collapsed, by
default the description is collapsed.

The expanding and collapsing causes a resize event not handled by the
pane causing a huge slowdown. This will needs some love.

The code is used to experiment with a different approach of the
implementation of a listbox.


Former-commit-id: e0fbeb71c9381e113e5db20bc6197dfe92e5473e
Former-commit-id: bcdc696001e7fb4a545dfbfc035201c7cac53e3a
Former-commit-id: e4b6821a7ce7ed1418e74855ab18702da34ce0b6
Former-commit-id: 70f7bb7343da2d0269755cde6313a328ed168d8d [formerly 15dbff0a7556773972256f00228823721fb240f7]
Former-commit-id: 4bd21e9c7eedbcb2d118fbc32b1f9d5fe4f4fb42