Add a debug and test code.
In order to test with formatting growing listboxes I need a growing
listbox. When showing the addon dialogue only one is shown, the others
are loaded with the load button. The code is a bit hacky, but hidden
under the --new-widgets switch.

The code is used to experiment with a different approach of the
implementation of a listbox.


Former-commit-id: decf68706383f56005477f8d1d5cd803bb452f77
Former-commit-id: c92b2ab8508c05364a38816375715ae46f01bb57
Former-commit-id: bdb2080d3fa90a2c94f52d964d40353d78fd6b8e
Former-commit-id: 0d40182af1452992e38f151489f7f572c61d60ca [formerly 49d02a3f435e677f4a69d67bed128fc0083dc882]
Former-commit-id: 329a1ca95c9e6b7553724ca7d9bb7437e4101c57