Fix the formatting for a growing pane.
The code seems to work for this specific case but needs more testing.

This commit finishes the proof-of-concept prototype it seems the big
hurdles in the listbox have been tackled. Now need to focus on getting a
real widget out of this prototype.

The code is used to experiment with a different approach of the
implementation of a listbox.


Former-commit-id: 2e22cd64696e1f397328c4bc60664433930b3e7a
Former-commit-id: 8dfac405478f0d9dd50f770a9c5ae19afd2747ea
Former-commit-id: 616d67eb590f2c60bc56cea972c35efe5b39cbf7
Former-commit-id: 9c8f4e44c864c550e677427873f879bb6368593d [formerly 4ca64d8458368aa6426ea933f914bfd444b7a591]
Former-commit-id: ef2c8325c27905b9ba9083b5446497beefcbc3d2