Move the widget fields to the widget class.
Some fields of the tbuilder_control are actually part of the twidget
class. These are now moved to the tbuilder_widget.

Some of these fields were also used in the tbuilder_grid and are moved
likewise. This means the tbuilder_grid now has some extra fields for
debugging. (`Inherited' from the tbuilder_control.)


Former-commit-id: 8da8a6d7484dbb5fe9c5d3c8ea6edc7c8bdbd09a
Former-commit-id: 1ab3956d8b7e70b69fe29667177fc774b5c9c65a
Former-commit-id: 30943c608d0362c490176195b9a688f175bc2812
Former-commit-id: 9f372f3f60a0f5bf759b6290da6b1f5a8c37bc7f [formerly 5ff41eaa49eac1aacfa13534de1dea9fc4ae7be5]
Former-commit-id: 91e17d93641be67c3c9376c51e7f28b617b0ec50