addon/mg: New add-on installation prompt for add-ons with VCS/.pbl files
Until now there was only a single, verbose prompt for the .pbl case.
Since people may use a VCS to contribute to an add-on without having a
.pbl file, that case is also now covered by this commit.

There's a single prompt with a list of sensible content that might be
"permanently lost" on installation. The list may contain VCS or .pbl
files for now, but perhaps at some point in the future someone might
want to make it into a full list of extra files or something.


Former-commit-id: 1931e7ed5923f321a34d27c80aa4d4a6ff0a4e75
Former-commit-id: 64bec3082410ed78f6d2367b10cf8ee8b9b4ba1f
Former-commit-id: 43d5e58a543b9f9702820130f67640ff46a86e2e
Former-commit-id: 29a529fcb284bea1aed195e9545dc73fdcff0889 [formerly 980eb0411b38024fdc67e0d9e05f2474a456a898]
Former-commit-id: 640c335040c4d8cbad1dc1d1e072a3608750a1c6