addon/mg: Rearrange list columns in roughly descending order of importance
With translations, running out of space due to the Type column and the
status line (which expands the Title column) is an inevitability. Thus
we need another compromise.

Type has to go last since it can be the longest due to the non-obvious
acronym usage (e.g. "MP scenario"). A later commit will introduce more
advanced filtering to make up for it when it goes missing due to space
constraints.

It should be noted that this is mostly relevant only for horizontal
resolutions < 1024.

The old column order:

  AM: Icon Name Version Author Type Downloads Size
  DP: Icon Name Version Author Type Size
  UA: Icon Name OldVersion NewVersion Author Size

The new order:

  AM: Icon Name Version Author Size Downloads Type
  DP: Icon Name Version Author Size Type
  UA: Icon Name OldVersion NewVersion Author Size

(AM being the Add-ons Manager; DP, the Dependency Prompt; and UA, the
Upgradable Add-ons view.)

The most important difference is that Size goes always before Downloads
and after Author now since people with low resolution displays are very
likely using services with poor/restricted Internet traffic; making sure
they can see the Size without reading the Description dialog seems only
logical.

An upcoming commit with make the gui2::taddon_description dialog display
all of the information just in case.


Former-commit-id: 7a0c4edcadc4481a2fe56bb19a30c1776700d999
Former-commit-id: df5a144eb46d9a8bb86106011aa847f6e5c22182
Former-commit-id: ce3b5b75a4a738642aa104ba36c87bf67fa10971
Former-commit-id: 394c8f4618587eede73546281fd71070ecdb6f07 [formerly 49407df41b4270f4cd8e339ddcd384be293a0eb7]
Former-commit-id: bb1c756070c18dbc35994ebc26313e1c5c0ca8d3