Add a new instance dummy widget.
This widget is used to postpone instantiation of a widget. This will be used
in the new matrix widget where the definition needs a widget which will only
be available when the matrix widget is instantiated in a window.

The code is used to experiment with a different approach of the
implementation of a listbox.


Former-commit-id: 5cb33d5bcb1243eb3601e61f3f915b442b9296fb
Former-commit-id: d1d05a48dab9c4fc7f6b6506f686b8dfc41df87d
Former-commit-id: 58d41c8f712d5b3bf8c5f5e314893b0d0bb39cb4
Former-commit-id: 2b477233a945584c3987af345ec79fbd4710621c [formerly 47d0ca9ff1180b8aba9e261315516a8ac714fe1a]
Former-commit-id: 717c085c40efe0d068c76061105f1bcac6343048