Switch the default gui2 render engine.
In gui2 there are two render engines; one using the widget's position to
render and the other based on the requested coordinates. The later is
used in the new widgets (pane, viewport and matrix) exclusively, meaning
these widgets don't work with the old engine.

The new one has been tested a while under --new-widgets and seems to
have no glitches so make it the default.

The old code hasn't been removed since doing so will force a huge
recompilation and it's desired to quickly switch back in case there
has been a regression.

The code is used to experiment with a different approach of the
implementation of a listbox.


Former-commit-id: c9897767cd4fbd1d88bd2911b1d6f1bbddf29d56
Former-commit-id: 2836bab47ba91322adb5d6f28b15d551ae026ad0
Former-commit-id: ef927cdaa12a0ef9f43775cd107e0d10923e78fc
Former-commit-id: f94a05a58ba958a4b1a85cb443a23bb427d27b1f [formerly 9c91ba8c1fe9920bfd4e8bb3a477fa56d89d6035]
Former-commit-id: 15dd9d8a99b546f3aea5a69ca1b23ae03ad8865b