addon/mg: Greatly simplify UI logic when no add-ons can be displayed
We now display the add-ons manager with an amount of controls
hidden/disabled and a line of text explaining the situation in a few
words.

This was necessary to reduce complexity since there are now more
situations to handle because of the add-on type filtering options. It
should look better this way, anyway, and should even be less annoying
for the users since the dialog no longer resets the view mode on its
own.


Former-commit-id: 2ab496447079e8030a4aa1a6a4b08f57365fd263
Former-commit-id: 70d5e34cd7987b2b7d30c3ea97c9c04aab47439f
Former-commit-id: c7cc5bc0fc074f987ebf781d439e39472891c2e5
Former-commit-id: aa3a9d2abbdef9ce98d60e7d4f8cb8fce5d5f526 [formerly 9056e3f82113ec8f0f2d5100ed671b981e488a27]
Former-commit-id: ce244f45bd2b578c69fc539392ce79b413d729fc