Adds the ENABLE_SHARED_LIBRARIES CMake option.
The use of shared libraries makes compilation debug versions faster but
results in extra shared libraries. For installation this is not
practical since the libraries need to be given SONAMES and kept in sync.
Therefore installation is not supported for this option and will
probably fail.

The statistics for the change are:

Static build:
  - Normal:
      real 20m51.604s
      user 19m47.838s
      sys  1m5.100s
      size 94MB
  - Debug:
      real 24m48.577s
      user 23m3.114s
      sys  1m22.305s
      size 1.9GB

Shared build:
  - Normal:
      real 21m5.141s
      user 19m56.879s
      sys  1m7.692s
      size 29MB
  - Debug:
      real 24m24.239s
      user 22m55.278s
      sys  1m20.193s
      size 596MB

So the build times don't change that much but the sizes do. When
rebuilding after changing a file the build times start to improve for
the shared build, but not timing done.


Former-commit-id: 61790f84e42735f367ca6fdae5457498c5df105e
Former-commit-id: 9232e63e3c6c8a609e88de26fbedb3ef86aaf2cb
Former-commit-id: 00de342f55a00e4213bdf5e034334ebbdcb08e8a
Former-commit-id: 572b0a486d320a54e837e41e1d4eb1ae4ee6aee1 [formerly 16eca76743f8450fb5da3e732c39b369bd5423e1]
Former-commit-id: 20d26a3bd71a52166da13630f3678d412dfaf5f8