Refectored move_unit()
This more object-oriented approach should be more robust in the face of
whatever future changes might come along. Plus, the code might be easier
to follow since the logic is broken into smaller pieces.

(There is some more streamlining to come. At this stage, very little
was changed with regards to the animation support that is required by
the enter/exit hex events.)


Former-commit-id: c296d4bf567d46ea4892ae48cff78838d6cffae5
Former-commit-id: cfce8021bfbac4bc8115a86e8a3c0d0eb1a0e88d
Former-commit-id: 222534655f9f350179de1f9cf5412cc02829754a
Former-commit-id: 128dea00434ef60d0d11dff68f6bfbadfd8664fc [formerly 599d91499de7f2fa64be863700fa4ff623b80060]
Former-commit-id: ab0ed248a625c4ad1b6517159dfe38c073a7aa9d