Fix game_display::fake_unit assignment.
One version left the fake unit in a bad state (not on a display, but recorded
as being on one), and the other failed to keep the fake unit on a display,
even though the comments indicated it did.

With this change, the display is preserved over assignment, which is consistent
with the copy constructor and the documentation for assigning a unit to a fake.
It is not easy to preserve the fake_unit's position in the deque, and currently
no benefit in trying to, so that part of the documentation has been changed.


Former-commit-id: a642724a89dd3a913c2069aef397ea43eaf0840f
Former-commit-id: 989fbb5c31b5c53ddb5dc0241d74ba54d9183dbd
Former-commit-id: 2043811e8e70d4cdbd36c95582c9e2d84d8e2ff4
Former-commit-id: 3c2a42dd4e208313df9d60548ab5d6f9c6f3131a [formerly ecc89f694caf7e844e349528bae535c1940e7873]
Former-commit-id: 8786297280a1bc5975678deff51958ff94e28e74