Do not always clear end_turn_ and hold_position_ in unit::set_movement().
This allows the function to be called without side-effects, so it can
be used as a simple setter (as the name suggests). Since the previous
behavior was to have side-effects (which is bad style for a default),
this commit introduces a non-optional parameter, ensuring that I've
looked at each place where this function is called.


Former-commit-id: f68e1c5205ab5a0a8d3905fd1e7482550b476c6b
Former-commit-id: f54aeedff3c3e3d4e458e391a5293625d7c74ce9
Former-commit-id: 5273bf1d3914bb8550e401fb4e3ff15b00a4fe5f
Former-commit-id: d833b6f58092456c0fc668e0815028debe3b0407 [formerly d48059a03852bd7616240a09ac0261d37fe8f2eb]
Former-commit-id: 7512b3f8a41f93b09938826c541c6700f8095d39