Implement [effect] apply_to=experience (bug #16747)
This implements 'increase' and 'set' attributes with similar semantics
to apply_to=hitpoints effects. The 'should_advance' attribute from the
FR is not implemented at this point since the unit class doesn't really
know how to present the user with an advancement prompt, and it
shouldn't really be its business to do so either.

As-is, using [effect] to set a unit's current XP above its maximum
simply won't trigger any advancements. It should be possible to work
around this limitation with [unstore_unit] (which is a real WML action,
unlike [effect]) and other tricks.


Former-commit-id: 822cb936d440ffd4575c12fd551a2a32c04c2571
Former-commit-id: a986a4b8a6daac272c2fb6dc2c772386acd7ad39
Former-commit-id: 82bc120a8eee72604577511973614d2e21387537
Former-commit-id: 670c10fd11ee92e7e9d3dc4ce0b73cfff1f957cf [formerly 2a56ae11680379e56f3c9295d1e84e25cc24e313]
Former-commit-id: 5bb897dee3f797192f261209e415bc7473db5124