Refactor weapon special contexts.
This allows some simplification both when the context is set and (in
one spot) when the context is used.

The semantics for the location parameters to set_specials_context() have
changed; attempts to use the old version will give a compiler error as the
function signatures have also changed (all current calls have been updated).


Former-commit-id: 3dba3d21d3a222c277db2c94c9aacfc0fba55f00
Former-commit-id: 709bdd34f2b1ddb68d8847c66b8e97e33cb06225
Former-commit-id: 4cb35d6b9b99ccdb28d593908b27c6226bb4f3aa
Former-commit-id: 424f2cb79d69ac49e0416c0a4889d699a59d1e64 [formerly 1b0fbf1072a93843cc9215906236370154d22cd6]
Former-commit-id: 72eab1e8301e7d4d09fc06819028da551b26b9d1