Correct calculation of swarm attacks when predicting a series of attacks.
(The conversion to double led to bad rounding, and the calculation should be
based on expected hit points, not original hit points.)

This is the last of the simple-to-fix bugs I detected with the stand-alone
attack_prediction / wesnoth-attack-sim combo. There is one remaining bug
(also related to swarm and a series of attacks) that might be a bear to rectify.


Former-commit-id: 9a9a7181dee5df2b3c414be341267348522fb306
Former-commit-id: ca48bbe5f5bc2449a25c9d6f5684205c8bccc84d
Former-commit-id: 0dd8595e2009695facd4b18d1bb34c30006723b2
Former-commit-id: 61c41ae2e6e79b32ede0df76729239b4e0945d81 [formerly f766c366106fd75ce4dd89c87c52ecf7232e669b]
Former-commit-id: 2ff21b9132c7fc39e3b32f4a2f85570f6e9fc1da