Correcting attack predictions for swarm...
...when there are previous combats to account for.

(It passes the attack simulator check.)

The basic problem with the old system is that it treated the number of
strikes as an independent variable, even though it depends on the
number of hit points.

There is some additional overhead with this method, but I optimized
some things for the common case (no swarm), and benchmarking indicated
a negligible 1% increase in run time when no units had swarm.


Former-commit-id: 86da939f9de2206c0c22c11c0bba23fd168d9764
Former-commit-id: 59d53a48f5074315d9addcfa554628988b666c1b
Former-commit-id: a93e7ff0ecac404f562eb8ad0af25200f6d59228
Former-commit-id: ad821ca5b7e27483d32d496ab6037f36893abde3 [formerly 58f1ec0ae2bd46cd95e2fe1420aee5d0d88c5b47]
Former-commit-id: 494e1066902ae21ad96a5a141243ea263ad3b5a9