Use size_t rather than unsigned long for the 'stored inside a pointer' type.
unsigned long is equal to void* on ILP32 (win32/linux32), LP32 (win16)
and LP64 (linux64) systems, but not on LLP64 systems, which win64 is
an example of.  http://www.unix.org/version2/whatsnew/lp64_wp.html

size_t is not guaranteed to be equal to void* either (see the 8086),
but uintptr_t is only provided by C99 and C++11, and even then is
optional.


Former-commit-id: 4865f06729f83dd1f6a826793968d506de3e6a70
Former-commit-id: cb75873fa27e59fb6e4e3fad904056236713cfeb
Former-commit-id: 2395c0fb5282908c054b36bc523c44dfd7c70e75
Former-commit-id: 05590b0df04c5fb4b1a2d99a379c3a40644e39cf [formerly 48b8d536cba74fac7e0bf7955106426cfb67e4a6]
Former-commit-id: c305cd113d06212296517d41eb96e6f28ff1dfc7