Instead of removing prestart events...
...(with a method that fails if an event's name is, e.g.,
"prestart,turn_end"), simply do not fire those events when loading a
replay (analogous to not firing those events when loading a game
normally).

This reverts most of 2012-11-07T19:28:44Z!therattigan@gmail.com, replacing it with a simpler fix to
bug #20272.  This also handles the additional case of start events.


Former-commit-id: c37e46ae563a682d9e600c0426b593dc28228bbd
Former-commit-id: c84809c8edb69ff947b85e45d422cc61d4783769
Former-commit-id: e8c692a32fd3d8ca2a79c2872feece234ff2f813
Former-commit-id: a34a7e5d3f5c70c3d23186f443c1eb406217de00 [formerly 14ba333ca93f416d5d2e5e9095dd410c00d333d9]
Former-commit-id: 194691eb8b42cae79b1201078e0ad667147a18c0