When creating [replay_start],
...have the snapshot overwrite the scenario config, rather than the
other way around.

This is important for the RNG as the snapshot includes how often the
RNG was invoked in prestart and start evnets, while the scenario's
config has the RNG recorded as having never been invoked (naturally
enough).  It might solve some other replay issues as well. (Not all
replay issues are well-documented.)


Former-commit-id: affd67fc42945b866c1482eec32a8ba1cf7012f7
Former-commit-id: 3bbc3003cd0a596353f929634cf72ae7ba9056ed
Former-commit-id: 54213758139d97c0773584e0260036adaedca5cb
Former-commit-id: df76b015472ea5a9ba5e55672ccdc737bcee0701 [formerly 736e8fdd4cadee6a1c32f9d2fb947bdf52df8293]
Former-commit-id: f7eda8abb0e06f94826fc25dd0f37c01e613c213