Restore 2010-07-20T10:46:41Z!guillaume.melquiond@gmail.com,
by removing 2010-07-23T05:51:17Z!upthorn@gmail.com 
and the follow-ups 2010-07-23T06:31:24Z!guillaume.melquiond@gmail.com and 2010-07-30T21:49:33Z!koraq@xs4all.nl.

In theory, the underlying bug has been fixed, so no need for the special
code for MSC. If this is not the case (the symptom would be "the windows
executable crashes immediately before loading the main menu"), feel free
to revert this.


Former-commit-id: a36f54f081cdfbec96458a85d2888da482d12826
Former-commit-id: ff62a725c867d91046bddca9622f6039f1570c7b
Former-commit-id: df439c816d9076c4c08a7f534d369a726ec7081b
Former-commit-id: 4e240c284e249f49dc87d6bbea30e8b3e2a8b196 [formerly aa9d2f9b48b074bc06ed4229594bb5911210b39d]
Former-commit-id: 909b3a4f45e002bf0e2116a4af1e25145023065a