Fix the presesent in-game command line autocompletion feature...
...so it actually works at all

Because of the questionable console commands handler design, the command
list isn't populated until the handler's dispatch() method is called for
the first time. We can call it with an empty command string to work
around this; the empty command is ignored and we get the command list
populated so we can actually build an autocompletion dictionary with it.
Previously, the autocompletion dictionary would be always empty.


Former-commit-id: 7d6cef4a13246c0d8ecb09d9d9a1bc36280ccea0
Former-commit-id: 24caf96c799a9bbe9e6f1fb5f3c3fc4957ee91f4
Former-commit-id: 28350930091a8f7ea44f35af58e8063a9f9939af
Former-commit-id: b9e1a5fc329045e1d7a2ebdd003b1cb267c28911 [formerly 651ae7a99c7a8f619204772559ea7da8f9b49100]
Former-commit-id: f386b05fb1245996e695c3dc60f29d02e53de098