gui1: Rework textbox event forwarding  to eliminate duplicate...
...(actually quadruplicate) events.  Changed from 2012-12-23T06:10:46Z!shadowm@wesnoth.org.

Due to the way focusing works, the focus switch in the middle of the
forwarding action would create a duplicate event for the origin textbox,
resulting in a duplicate event for the target (a total of four events to
process).

Instead, make the implementation of gui::textbox::handle_event()
forwarding-aware, and provide a wrapper for the events engine
implementation to use by default.


Former-commit-id: 62eab4a5de0b0115fdecd75433b8c23632a37fa3
Former-commit-id: 04c38a056edb2d531c5b60937f03f5c943fae837
Former-commit-id: 369edd39d7f039edae6ee13d2723054f5e8a2037
Former-commit-id: 9901ff2d37d2330b7fbd5b3e39e2f4c712140d2d [formerly 9c73533e82cece9be84b0d816255f12ad25f9c42]
Former-commit-id: ad27a3e0fec2b8e265a924cbedfc9fb71719902e