Make unit_type_data::build_unit_type() public,
...and update various parts of the code to call it (instead of using
unit_type_data::find() to "find" a known unit_type).

There was one call in the editor's code where I am unsure if there are
supposed to be side effects other than building the unit_type. So I
left the code as-is and just added a comment for someone else to
evaluate later.


Former-commit-id: 1715f859038c6a04ede1f401dba7ba7fc28311ef
Former-commit-id: 644ba94dd73bccc04efe6af2cfda0a4f0d201e27
Former-commit-id: 4dfea9ba779df86667621def2b1525afbfe7627f
Former-commit-id: 3897c42201d91fbf9918f17f07f9c54042fc81ce [formerly b2a3a18092123127069774c32a61eac4484599fe]
Former-commit-id: ae0a12bce9772fa5575fe8341178e4345ebc9837