Set up multiplayer games started from the commandline...
...equivalently to lobby games.

This means that parameters now use the correct default values,
scenarios using random maps work and side leaders always have unique
IDs.  Scenario and side parameters are defined as follows, in order of
increasing priority:

  Wesnoth defaults --> map settings --> commandline-specified values

All modifications are in a code path separate from lobby games (as it
was before), so lobby games are not affected.

This fixes bugs #19853, #19883, #19895 and #20009.


Former-commit-id: 425f9455edf14619d43a3e4c1150280e517154e6
Former-commit-id: 85e9ae81fd4b9aa24d14a659c02d479eeb20bc08
Former-commit-id: 9fae18a252c610bdde18eb6240f8d781e123f667
Former-commit-id: 26bc94f2833b0b505d7cdf5ee9d98d3e55ab039c [formerly 16ff8e6887685daf3cc43106ba25dd901b120bb5]
Former-commit-id: 865d8beb02f209f8b35929f11b019ea2f0c7e5b6