Move processing of child configs from unit_type to unit_type_data::set_config().
It is a bit hacky to have unit_type modify the config passed as a
parameter to its constructor, especially when the modifications are
not done by the constructor.

As for performance, this change has little impact on the net time
taken by set_config(). While it does spend more time processing
configs, it takes less time building unit_types. The net result might
be a small increase in time, but any increase is comparable to the
variations in time seen amongst multiple test runs.


Former-commit-id: 555e9ac8f0f0b95595bd83e545ef8d50bebc736c
Former-commit-id: 10a0ae311330044bee185a463684bd22b39fa41e
Former-commit-id: 835a2654c38e9c1df93240ad9b340bfe7bfd5c52
Former-commit-id: 97288f80df0c7b801607287c107377771ae86a80 [formerly 49ce586c20da637089d312f7493704a05e9a75d6]
Former-commit-id: 1f692eeb458cc08a773cf5ed31d753e1136de62c