Add new AI aspect leader_ignores_keep.
If set, AI leaders do not move to the closest keep at the beginning of
the turn.  Instead, they participate in the move_to_targets candidate
action (and all other CAs in which they already participated anyway,
of course).  The default value is 'no', which leaves the default
behavior unchanged.


Former-commit-id: 74ef541fc099c9b2e8b9502e3cc399053943ff0b
Former-commit-id: 4646a25731af05b81c1139e8d74e5517480941d9
Former-commit-id: cedc1756c6996b8b1e8331ddfd02bea242ea3d97
Former-commit-id: 201a627c7d646e9d44c062ecd43798c56a7fb688 [formerly a2cb82ad175834acd88c297a563eac5a480ba05f]
Former-commit-id: e4d07674e7f83da25fa91b9fd4c871f7b396bd6b