Avoid click dismissing moving a unit.
When click dismissing a dialogue in the past the DOWN event was used.
This lead to a bug. The obvious change was to switch to the UP event,
this lead to another bug; the dialogue was directly dismissed. Since
the game map code uses the UP and DOWN event to select a unit there is
no simple solution.

Upon entry this value stores the mouse button state at entry.  When a
button is DOWN and goes UP that button does _not_ trigger a dismissal
of the dialogue, instead that button's down state is removed from this
variable. Therefore the next UP event does dismiss the dialogue.

(Fixes bug #18970)


Former-commit-id: d2fa8fc293a8b95fe0172a9e7c6f282717d55439
Former-commit-id: 822e9a2be322d8fd8e13ca027d2bc223787c06c3
Former-commit-id: 00f80bfcc043d956dc67fbd78a0064fd100d9ddb
Former-commit-id: 9344cc5ac1534b328057e675dfd39eb213dfdb49 [formerly 13b4e8f9e20b164eaa82e978c61c39fa8ac1b877]
Former-commit-id: 97c0782a150897e8a4efe5de2b9258203f67be11