Move halo clearing from unit_map.cpp to lua.cpp.
Apparently, halos were not getting redrawn when the AI moves units
around for its calculations. Instead of adding a bunch of not really
necessary invalidations (to get the halos redrawn), restrict halo
clearing to a case where it is needed (Lua's extract_unit, which is
used by [move_unit]).

Also add an invalidation to Lua's put_unit, in case someone's Lua code
attempts to do something similar to the C++ AI.

This replaces r56143.


Former-commit-id: c3b704d56cac0ec8dfc5bad873962b855d711ec3
Former-commit-id: 1bb232ed456df09c0065f433c3b9563bf9e3cf66
Former-commit-id: a6c8b593583efaa54f22a2c20a0b1d161441a5ff
Former-commit-id: 3d9d19a31afdb7f42ce2c99a0ba246a4060ae577 [formerly 3523e3daf1f93672364cffe7178e60f4a5d016ba]
Former-commit-id: eb37636a4ff82f3b241f50467bedfb7cc98508aa