Move halo clearing from unit_map.cpp to lua.cpp.
Apparently, halos were not getting redrawn when the AI moves units
around for its calculations. Instead of adding a bunch of not really
necessary invalidations (to get the halos redrawn), restrict halo
clearing to a case where it is needed (Lua's extract_unit, which is
used by [move_unit]).                                               

Also add an invalidation to Lua's put_unit, in case someone's Lua code
attempts to do something similar to the C++ AI.

This replaces r56340.


Former-commit-id: 9dcca8346e02705699bb0f26f4976e7fe3c2b49a
Former-commit-id: d5005ea1ba34cf05f1bef52e2689721871fd410f
Former-commit-id: 3d5982238f544ac1ffe6a9aeeded3d66951a563e
Former-commit-id: 180256b4d6ba3b4b9566a7c97f0c54444929c9d0 [formerly b12c9ffdf2f48f954bda01739c3679b1c03f8563]
Former-commit-id: 692698acdf901aea807a8238118de2d0cafbe2be