Reduce the data stored in the undo stack.
Most actions had caused a unit to be stored in the stack, but the stack
does not really need all the data from a unit. Now it will store only
those parts it needs, which might have a significant impact in
scenarios with WML-heavy units. There is also a reduction in the size
of mid-turn saves (~9% in one test case using NR:Showdown).


Former-commit-id: 07529a13c728843171d44d8900c8fe7edf6bd2e0
Former-commit-id: 24d9468c07fa45d77a598a8b7ee70fd75664f921
Former-commit-id: d983d5f97ea1214eae0791cce48d38585922a3a1
Former-commit-id: 4500147265a794b18d29c157414188225a60017b [formerly c77a59c3b87a17f5803381c8a7ec181905fc91c4]
Former-commit-id: 86295d7068dd86b30a0c1ce6e2455a0ef1f8f58f