Polmorphism for undo actions
The old implementation (a single struct covering all types of actions)
was getting a bit over-extended with the evolution of the kinds of data
being stored. This new approach derives a class for each type of action
and relies on virtual functions to decide what to do.

To support the polymorphism, undo_list's std:vectors have been changed
to boost::ptr_vectors. This has the beneficial side-effect of changing
several struct copies to pointer copies.


Former-commit-id: b0f8d5b45646e4e3bc09234598f6ea3730a4f200
Former-commit-id: 013c98ced6e70bda363a4866b43f3de56d9d10c8
Former-commit-id: 73a4641dc60295f9ed2aa9d4df7f92d5bafe793e
Former-commit-id: 3002143094f10ea586f305b2cca7e5c5d838ea79 [formerly d8289784c288b3d20d1a3d4eca622d69c129ebe4]
Former-commit-id: f47f72f06b7bd5fcdf8c20ef6b1cbaaa6c2f18ec