display: Fix broken conditionals in display::scroll_to_tyles()
(Backport note: yes, I typoed that; it should be
display::scroll_to_tiles() in the commit message.)

This purportedly fixes bug #18793.
(Candidate for 1.10, PLEASE TEST!)

The conditionals in question:

> if (scroll_type == ONSCREEN || ONSCREEN_WARP) {

This *always* evaluates to true when scroll_type != ONSCREEN, because
ONSCREEN_WARP is an enum member with a non-zero value. The bogus code is
effectively the same as:

> if ((scroll_type == ONSCREEN) || true) {

Which was most likely not the author's intention (fendrin in r49988).
Therefore, I've replaced the conditionals with what should be the
correct way to do what I suspect was meant above:

> if (scroll_type == ONSCREEN || scroll_type == ONSCREEN_WARP) {

This seems to fix the viewport centering issues for me, but it would be
nice if people reported their own experience with the fix before
backporting it to 1.10.


Former-commit-id: 983e68d2900f41440f9a256e4ff0925a33c1c8d1
Former-commit-id: 1f86e4ae8f9f4b535472fe4a8dc14bbf67c0e0ff
Former-commit-id: 670a4c5031e4bdc561d380e752decdca1653abf5
Former-commit-id: 6f985583e1549c15b36ab06f4ac967c96f45c211 [formerly 723012f8763cd19fc207352b61c659d2f15be396]
Former-commit-id: 8c72904ca3862473b2cd2039da38f0a593ef9c25