display: Fix broken conditionals in display::scroll_to_tyles()
This purportedly fixes bug #18793.
(Candidate for 1.10, PLEASE TEST!)

The conditionals in question:

> if (scroll_type == ONSCREEN || ONSCREEN_WARP) {

This *always* evaluates to true when scroll_type != ONSCREEN, because
ONSCREEN_WARP is an enum member with a non-zero value. The bogus code is
effectively the same as:

> if ((scroll_type == ONSCREEN) || true) {

Which was most likely not the author's intention (fendrin in r49988).
Therefore, I've replaced the conditionals with what should be the
correct way to do what I suspect was meant above:

> if (scroll_type == ONSCREEN || scroll_type == ONSCREEN_WARP) {

This seems to fix the viewport centering issues for me, but it would be
nice if people reported their own experience with the fix before
backporting it to 1.10.


Former-commit-id: 40271734b045d1706abd8e4fd47f738a9faf72f2
Former-commit-id: b04bab7b984dd7416e2145de2224d5c516d609df
Former-commit-id: ee1b608217b70196b34c6854827294730a4110a9
Former-commit-id: 033da902343eedbbcd32f1934b44eb5bcf06d7fc [formerly 567f3b0f4b2aa53343e77a6e4a87b1c24a99dfc1]
Former-commit-id: 748aa1f074480561b1cdaf3490d91068b5d01d2a