remove '~' in [textdomain] and [binary_path] paths
Although the original purpose of the in_textdomain and in_binary_path code, an aborted effort to update their paths to "data/add-ons/", has been superseded by code that updates those paths on all lines, it can still be put to use.

Our first step is to move that section below the code that updated UMC paths, so our regexes won't have to deal with "@campaigns" and "data/campaigns" strings. Then we delete the 'if 0:' line that was neutralizing this section, as well as the obsolete path-changing code. The rest is de-indented one level.

Then we look for the use of "~" for userdata, which does not work for textdomains and binary paths.

Our regex object, 'tilde', is constituted thusly: (1) We make sure that the line starts with the "path" key. Any line we're interested in ought to start with this, and this will also keep this code from going wild on the campaign includes, if an author forgot a closing tag (no reset to False). (x) There shouldn't be any whitespace around the = sign, but we'll be kind. (2) On the value side, there shouldn't be anything before the tilde except perhaps a quote. Rather than underestimate the ingenuity of authors in coming up with weird code, however, I allow anything except a comment to match for a few characters. But if we haven't hit the '~' after five characters, I figure something's wrong, and bail. (3) Then we come to the tilde. Normally, it would be adjoining "add-ons/", but some authors interpolate a slash, or 'data/' (here represented as an optional string).

If we match, we rebuild the line, except 'data/add-ons' is substituted for group(3), and we log to stdout.


Former-commit-id: d79c3ff9ed1f3730d260e77ee6e5e9658e08c0a9
Former-commit-id: 3334a9c8805102ee38849402c8513b973e4a3aa4
Former-commit-id: b0382799a164fd5544005906ca317e7d52b860e6
Former-commit-id: df63be493c532f1ccabdbd6adc74384c1f88b28c [formerly cbabf97ef8f153e62fb1340d7bb16ea73482765a]
Former-commit-id: fcdbfaa9257ac9c4812bc16fd575bac7240998a2