Fixes the text box cursor position.
The placement of the cursor sometimes was wrong under Windows. (The bug
could be reproduced under Linux by changing the configuration files.)

The problem was caused by the height of the text box and the permission
to wrap the text on a new line. This caused the cursor position to
return to the beginning of the line.

Fixes bug #17220 using patch #3822.


Former-commit-id: 65d1e7e0b07959620ecd022b7150e847be6ab4f6
Former-commit-id: fb886dc2ccde90024edab43c11f9b195c026857e
Former-commit-id: 5d40a84b8aa20d53d90dc2a6d9d177b4d26fe124
Former-commit-id: 3dde41e6b99182f16eb1a2be3ffebe5af9f3672f [formerly 279d224717592b4079552bf732784b595118f1f7]
Former-commit-id: 57d72f83e848e7f30bde65d539ed31f8d5f659e1