Fixes bug #20765: Stack overflow in animation because of weird WML files.
A call to unit::redraw_unit would check if the unit contains any animation. If
no animation is found, then it would call set_standing to put a new animation.

States are usually changed through unit::start_animation. unit::set_standing
uses this approach also. If unit::start_animation is unable to change the unit
to the animation passed as argument, it would try to animate the object calling
unit::set_standing (infinite recursion).

unit::start_animation had a check (maybe to prevent the infinite recursion), but
this check was buggy (if the intent was to prevent the infinite recursion).

The solution add a instruction to successfully set the unit to the state
STATE_STANDING and fixes the check that was using the wrong logical operator.


Former-commit-id: 02709b5a517a8ec61a61d554b5ae6fb8de6bf23f
Former-commit-id: 37c2f6265cda5a8e93f3d8ccf105450928003dc6
Former-commit-id: b9f2b893277038bf332c800975d8010f2494c2be
Former-commit-id: 99e960c20a8c4263675ac8a03e46d5212eebd039 [formerly d0a572ad2683f9ccd608db4ff1ff66f0cfcf13d8]
Former-commit-id: d6b43743036d81507c53e9120735c1bd8a911b30